% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskyr_class.R
\name{as.table.riskyr}
\alias{as.table.riskyr}
\title{Create table of risk information.}
\format{A "riskyr" object.}
\usage{
\method{as.table}{riskyr}(x, sep = " = ")
}
\arguments{
\item{x}{An object of class "riskyr", usually a result of a call to \code{riskyr}.}

\item{sep}{By which character should the entries be separated from the labels? (default " = ").}
}
\description{
\code{as.table.riskyr} provides a \code{as.table} method for objects of class "riskyr" to allow for readable output.
}
\examples{
table(scenarios$n4)

}
\seealso{
Other summary functions: \code{\link{summary.riskyr}}
}
\concept{summary functions}
