% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA1}
\alias{BRCA1}
\title{Cumulative risk of breast cancer in women with the BRCA1 mutation.}
\format{
A data frame (11 x 2).

\code{x}: age (in years).

\code{y}: cumulative risk of developing breast
cancer in this (BRCA1) population.
}
\usage{
BRCA1
}
\description{
\code{BRCA1} provides the cumulative risk of breast cancer
in a population of women with the BRCA1 mutation
as a function of their age (in years).
}
\seealso{
\code{\link{plot_crisk}} plots cumulative risk curves.

Other datasets: 
\code{\link{BRCA1_mam}},
\code{\link{BRCA1_ova}},
\code{\link{BRCA2_mam}},
\code{\link{BRCA2_ova}},
\code{\link{BRCA2}},
\code{\link{df_scenarios}},
\code{\link{t_A}},
\code{\link{t_B}},
\code{\link{t_I}}
}
\concept{datasets}
\keyword{datasets}
