% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA2}
\alias{BRCA2}
\title{Cumulative risk of breast cancer in women with the BRCA2 mutation.}
\format{
A data frame (11 x 2).

\code{x}: age (in years).

\code{y}: cumulative risk of developing breast
cancer in this (BRCA2) population.
}
\usage{
BRCA2
}
\description{
\code{BRCA2} provides the cumulative risk of breast cancer
in a population of women with the BRCA2 mutation
as a function of their age (in years).
}
\seealso{
\code{\link{plot_crisk}} plots cumulative risk curves.

Other datasets: 
\code{\link{BRCA1}},
\code{\link{df_scenarios}}
}
\concept{datasets}
\keyword{datasets}
