% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\name{comp_fart}
\alias{comp_fart}
\title{Compute a decision's false alarm rate from its specificity.}
\usage{
comp_fart(spec)
}
\arguments{
\item{spec}{A specificity given as a probability
(i.e., a numeric value in the range from 0 to 1).}
}
\value{
A false alarm rate as a probability
(i.e., a numeric value in the range from 0 to 1).
}
\description{
\code{comp_fart} is a conversion function that takes a specificity \code{spec}
-- given as a probability (i.e., a numeric value in the range from 0 to 1) --
as its input, and returns the corresponding false alarm rate \code{fart}
-- also as a probability -- as its output.
}
\details{
Specificity and false alarm rate are both features of the decision process
(e.g., a diagnostic test).
The false alarm rate and specificity are complements (i.e., \code{fart = 1 - spec}).

\code{comp_fart} is complementary to the conversion function
\code{\link{comp_spec}}.
}
\examples{
comp_fart(2)                 # => NA + Warning that 2 is not in 0 to 1 range
comp_fart(1/3)               # => 0.6666667
comp_fart(comp_spec(0.123))  # => 0.123
}
\seealso{
\code{\link{is_complement}} to verify numeric complements;
\code{\link{as_pc}} to display a probability as a percentage

Other conversion functions: \code{\link{comp_complement}},
  \code{\link{comp_spec}}
}
