% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_freq_freq.R
\name{comp_freq_freq}
\alias{comp_freq_freq}
\title{Compute full set of frequencies from essential frequencies.}
\usage{
comp_freq_freq(hi = freq$hi, mi = freq$mi, fa = freq$fa, cr = freq$cr)
}
\arguments{
\item{hi}{The number true positives, or hits \code{\link{hi}}}

\item{mi}{The number false negatives, or misses \code{\link{mi}}}

\item{fa}{The number false positives, or false alarms \code{\link{fa}}}

\item{cr}{The number true negatives, or correct rejections \code{\link{cr}}}

\item{N}{The population size \code{\link{N}} (not used yet, use for scaling to new population sizes)}
}
\description{
\code{comp_freq_freq} computes all frequencies
contained in \code{\link{freq}} from 4 essential frequencies
(\code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, \code{\link{cr}}).
}
\details{
\code{comp_freq_freq} is an analog to \code{\link{comp_prob_freq}}
that computes all \emph{probabilities} contained in \code{\link{prob}}
from the same 4 essential frequencies.

The following relationships hold (and are used in computations):

\enumerate{

\item prevalence \code{\link{prev}}:

\code{\link{prev} = \link{cond.true}/\link{N}  =  (\link{hi} + \link{mi}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item sensitivity \code{\link{sens}}:

\code{\link{sens} = \link{hi}/\link{cond.true}  =  \link{hi} / (\link{hi} + \link{mi})  =  (1 - \link{mirt})}

\item miss rate \code{\link{mirt}}:

\code{\link{mirt} = \link{mi}/\link{cond.true}  =  \link{mi} / (\link{hi} + \link{mi})  =  (1 - \link{sens})}

\item specificity \code{\link{spec}}:

\code{\link{spec} = \link{cr}/\link{cond.false}  =  \link{cr} / (\link{fa} + \link{cr})  =  (1 - \link{fart})}

\item false alarm rate \code{\link{fart}}:

\code{\link{fart} = \link{fa}/\link{cond.false}  =  \link{fa} / (\link{fa} + \link{cr})  =  (1 - \link{spec})}



\item proportion of positive decisions \code{\link{ppod}}:

\code{\link{ppod} = \link{dec.pos}/\link{N}  =  (\link{hi} + \link{fa}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item positive predictive value \code{\link{PPV}}:

\code{\link{PPV} = \link{hi}/\link{dec.pos}  =  \link{hi} / (\link{hi} + \link{fa})  =  (1 - \link{FDR})}

\item negative predictive value \code{\link{NPV}}:

\code{\link{NPV} = \link{cr}/\link{dec.neg}  =  \link{cr} / (\link{mi} + \link{cr})  =  (1 - \link{FOR})}

\item false detection rate \code{\link{FDR}}:

\code{\link{FDR} = \link{fa}/\link{dec.pos}  =  \link{fa} / (\link{hi} + \link{fa})  =  (1 - \link{PPV})}

\item false omission rate \code{\link{FOR}}:

\code{\link{FOR} = \link{mi}/\link{dec.neg}  =  \link{mi} / (\link{mi} + \link{cr})  =  (1 - \link{NPV})}

}
}
\examples{
## Basics:
comp_freq_freq()
all.equal(freq, comp_freq_freq())  # => should be TRUE


## Circular chain:
# 1. Current numeric parameters:
num

# 2. Compute all 10 probabilities in prob (from essential probabilities):
prob <- comp_prob()
prob

# 3. Compute 9 frequencies in freq from probabilities:
freq <- comp_freq(round = FALSE)   # no rounding (to obtain same probabilities later)
freq

# 4. Compute 9 frequencies AGAIN (but now from frequencies):
freq_freq <- comp_freq_freq()

# 5. Check equality of results (steps 2. and 4.):
all.equal(freq, freq_freq)  # => should be TRUE!


}
\seealso{
\code{\link{comp_prob_freq}} computes full probability information from essential frequencies;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{is_freq}} verifies frequency inputs.

Other functions computing frequencies: \code{\link{comp_freq}},
  \code{\link{comp_min_N}}, \code{\link{trans_prob}}
}
