% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_xxxx_prob.R
\name{comp_freq_prob}
\alias{comp_freq_prob}
\title{Compute frequencies from (3 essential) probabilities.}
\usage{
comp_freq_prob(prev = prob$prev, sens = prob$sens, mirt = NA,
  spec = prob$spec, fart = NA, tol = 0.01, N = freq$N, round = TRUE)
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{tol}{A numeric tolerance value for \code{\link{is_complement}}.
Default: \code{tol = .01}.}

\item{N}{The number of individuals in the population.
If \code{\link{N}} is unknown (\code{NA}),
a suitable minimum value is computed by \code{\link{comp_min_N}}.}

\item{round}{A Boolean value that determines whether frequencies are
rounded to the nearest integer. Default: \code{round = TRUE}.}
}
\value{
A list \code{\link{freq}} containing 9 frequency values.
}
\description{
\code{comp_freq_prob} computes current frequency information
from a sufficient and valid set of 3 essential probabilities
(\code{\link{prev}}, and
\code{\link{sens}} or its complement \code{\link{mirt}}, and
\code{\link{spec}} or its complement \code{\link{fart}}).
It returns a list of 9 frequencies \code{\link{freq}}
for a population of \code{\link{N}} individuals
as its output.
}
\details{
By default, the values of \code{\link{prev}}, \code{\link{sens}},
and \code{\link{spec}} are initialized to the probability information
currently contained in \code{\link{prob}}.

\code{comp_freq_prob} is a wrapper function for the more basic
function \code{\link{comp_freq}}, which only accepts the
3 essential probabilities
(\code{\link{prev}}, \code{\link{sens}}, \code{\link{spec}})
as inputs.

When using \code{comp_freq_prob} with the arguments
\code{\link{mirt}} and \code{\link{fart}}, their complements
\code{\link{sens}} and \code{\link{spec}} must either be
valid complements (as in \code{\link{is_complement}}) or
\code{NA}.

The value of \code{\link{N}} uses the frequency information
currently contained in \code{\link{freq}} as its default.
If \code{\link{N}} is unknown (\code{NA}),
a suitable minimum value can be computed by \code{\link{comp_min_N}}.

In addition to \code{\link{prev}}, both
\code{\link{sens}} and \code{\link{spec}} are necessary arguments.
If only their complements \code{\link{mirt}} or \code{\link{fart}}
are known, first use \code{\link{comp_complement}},
\code{\link{comp_comp_pair}}, or \code{\link{comp_complete_prob_set}}
to obtain the 3 essential probabilities.

By default, \code{comp_freq_prob} and \code{\link{comp_freq}}
round frequencies to nearest integers to avoid decimal values in
\code{\link{freq}}. Using the option \code{round = FALSE}
turns off rounding.

Key relationships:

\itemize{

\item Other functions translating between representational formats:

\enumerate{

\item \code{comp_freq_prob} (defined here) is
   a wrapper function for \code{\link{comp_freq}} and
   an analog to 3 other format conversion functions:

\item \code{\link{comp_freq_freq}} computes
   current \emph{frequency} information contained in \code{\link{freq}}
   from 4 essential frequencies
   (\code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, \code{\link{cr}}).

\item \code{\link{comp_prob_freq}} computes
   current \emph{probability} information contained in \code{\link{prob}}
   from 4 essential frequencies
   (\code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, \code{\link{cr}}).

\item \code{\link{comp_prob_prob}} computes
   current \emph{probability} information contained in \code{\link{prob}}
   from 3 essential probabilities
   (\code{\link{prev}}, \code{\link{sens}}, \code{\link{spec}}).

}

\item Two perspectives:

A population of \code{\link{N}} individuals can be split into 2 subsets
in 2 different ways:

\enumerate{

\item by condition:

The frequency \code{\link{cond.true}} depends on the prevalence \code{\link{prev}}
   and
   the frequency \code{\link{cond.false}} depends on the prevalence's complement \code{1 - \link{prev}}.

\item by decision:

The frequency \code{\link{dec.pos}} depends on the proportion of positive decisions \code{\link{ppod}}
   and
   the frequency \code{\link{dec.neg}} depends on the proportion of negative decisions \code{1 - \link{ppod}}.

}

The population size \code{\link{N}} is a free parameter (independent of the
essential probabilities \code{\link{prev}}, \code{\link{sens}}, and \code{\link{spec}}).

If \code{\link{N}} is unknown (\code{NA}), a suitable minimum value can be computed by \code{\link{comp_min_N}}.

\item Combinations of frequencies:

In a population of size \code{\link{N}} the following relationships hold:

\enumerate{

\item \code{\link{N} = \link{cond.true} + \link{cond.false} = (\link{hi} + \link{mi}) + (\link{fa} + \link{cr})} (by condition)

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg} = (\link{hi} + \link{fa}) + (\link{mi} + \link{cr})} (by decision)

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}} (by condition x decision)

}

The two perspectives (by condition vs. by decision) combine the 4 essential frequencies
  (i.e., \code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, \code{\link{cr}})
  in 2 different ways.

\item Defining probabilities in terms of frequencies:

Probabilities \emph{are} -- determine, describe, or are defined as -- the relationships between frequencies.
Thus, they can be computed as ratios between frequencies:

\enumerate{

\item prevalence \code{\link{prev}}:

\code{\link{prev} = \link{cond.true}/\link{N}  =  (\link{hi} + \link{mi}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item sensitivity \code{\link{sens}}:

\code{\link{sens} = \link{hi}/\link{cond.true}  =  \link{hi} / (\link{hi} + \link{mi})  =  (1 - \link{mirt})}

\item miss rate \code{\link{mirt}}:

\code{\link{mirt} = \link{mi}/\link{cond.true}  =  \link{mi} / (\link{hi} + \link{mi})  =  (1 - \link{sens})}

\item specificity \code{\link{spec}}:

\code{\link{spec} = \link{cr}/\link{cond.false}  =  \link{cr} / (\link{fa} + \link{cr})  =  (1 - \link{fart})}

\item false alarm rate \code{\link{fart}}:

\code{\link{fart} = \link{fa}/\link{cond.false}  =  \link{fa} / (\link{fa} + \link{cr})  =  (1 - \link{spec})}



\item proportion of positive decisions \code{\link{ppod}}:

\code{\link{ppod} = \link{dec.pos}/\link{N}  =  (\link{hi} + \link{fa}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item positive predictive value \code{\link{PPV}}:

\code{\link{PPV} = \link{hi}/\link{dec.pos}  =  \link{hi} / (\link{hi} + \link{fa})  =  (1 - \link{FDR})}

\item negative predictive value \code{\link{NPV}}:

\code{\link{NPV} = \link{cr}/\link{dec.neg}  =  \link{cr} / (\link{mi} + \link{cr})  =  (1 - \link{FOR})}

\item false detection rate \code{\link{FDR}}:

\code{\link{FDR} = \link{fa}/\link{dec.pos}  =  \link{fa} / (\link{hi} + \link{fa})  =  (1 - \link{PPV})}

\item false omission rate \code{\link{FOR}}:

\code{\link{FOR} = \link{mi}/\link{dec.neg}  =  \link{mi} / (\link{mi} + \link{cr})  =  (1 - \link{NPV})}

}

}
}
\examples{
# Basics:
comp_freq_prob(prev = .1, sens = .9, spec = .8, N = 100)           # => ok: hi = 9, ... cr = 72.
# Same case with complements (using NAs to prevent defaults):
comp_freq_prob(prev = .1, sens = NA, mirt = .1, spec = NA, fart = .2, N = 100)  # => same result

comp_freq_prob()                   # => ok, using probability info currently contained in prob
length(comp_freq_prob())           # => a list containing 9 frequencies
all.equal(freq, comp_freq_prob())  # => TRUE, unless prob has been changed after computing freq
freq <- comp_freq_prob()           # => computes frequencies and stores them in freq

# Ways to work:
comp_freq_prob(prev = 1, sens = 1, spec = 1, N = 101)  # => ok + warning: N hits (TP)

# Same case with complements (using NAs to prevent defaults):
comp_freq_prob(prev = 1, sens = NA, mirt = 0, spec = NA, fart = 0, N = 101)

comp_freq_prob(prev = 1, sens = 1, spec = 0, N = 102)  # => ok + warning: N hits (TP)
comp_freq_prob(prev = 1, sens = 0, spec = 1, N = 103)  # => ok + warning: N misses (FN)
comp_freq_prob(prev = 1, sens = 0, spec = 0, N = 104)  # => ok + warning: N misses (FN)
comp_freq_prob(prev = 0, sens = 1, spec = 1, N = 105)  # => ok + warning: N correct rejections (TN)

comp_freq_prob(prev = 0, sens = 1, spec = 0, N = 106)  # => ok + warning: N false alarms (FP)

# Same case with complements (using NAs to prevent defaults):
comp_freq_prob(prev = 0, sens = NA, mirt = 0, spec = NA, fart = 1, N = 106)  # => ok + warning: N false alarms (FP)

# Watch out for:
comp_freq_prob(prev = 1, sens = 1, spec = 1, N = NA)  # => ok + additional warning that N = 1 was computed
comp_freq_prob(prev = 1, sens = 1, spec = 1, N =  0)  # => ok, but all 0 + warning (extreme case: N hits, NPV = NaN)
comp_freq_prob(prev = .5, sens = .5, spec = .5, N = 10, round = TRUE)   # => ok, but all rounded (increasing errors: mi and fa)
comp_freq_prob(prev = .5, sens = .5, spec = .5, N = 10, round = FALSE)  # => ok, but not rounded

# Ways to fail:
comp_freq_prob(prev = NA,  sens = 1, spec = 1,  100)   # => NAs + no warning (prev NA)
comp_freq_prob(prev = 1,  sens = NA, spec = 1,  100)   # => NAs + no warning (sens NA)
comp_freq_prob(prev = 1,  sens = 1,  spec = NA, 100)   # => NAs + no warning (spec NA)
comp_freq_prob(prev = 8,  sens = 1,  spec = 1,  100)   # => NAs + warning (prev beyond range)
comp_freq_prob(prev = 1,  sens = 8,  spec = 1,  100)   # => NAs + warning (sens & spec beyond range)


}
\seealso{
\code{\link{comp_freq_freq}} computes current frequency information from (4 essential) frequencies;
\code{\link{comp_prob_freq}} computes current probability information from (4 essential) frequencies;
\code{\link{comp_prob_prob}} computes current probability information from (3 essential) probabilities;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{comp_complement}} computes a probability's complement;
\code{\link{comp_comp_pair}} computes pairs of complements;
\code{\link{comp_complete_prob_set}} completes valid sets of probabilities;
\code{\link{comp_min_N}} computes a suitable population size \code{\link{N}} (if missing).

Other functions computing frequencies: \code{\link{comp_freq_freq}},
  \code{\link{comp_freq}}, \code{\link{comp_min_N}},
  \code{\link{comp_prob_prob}}

Other format conversion functions: \code{\link{comp_freq_freq}},
  \code{\link{comp_prob_freq}},
  \code{\link{comp_prob_prob}}
}
