% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_prob_prob.R
\name{comp_mirt}
\alias{comp_mirt}
\title{Compute a decision's miss rate from its sensitivity.}
\usage{
comp_mirt(sens)
}
\arguments{
\item{sens}{The decision's sensitivity \code{\link{sens}}.}
}
\value{
The decision's miss rate \code{\link{mirt}}.
}
\description{
\code{comp_mirt} is a conversion function that takes a sensitivity \code{\link{sens}}
-- given as a probability (i.e., a numeric value in the range from 0 to 1) --
as its input, and returns the corresponding miss rate \code{\link{mirt}}
-- also as a probability -- as its output.
}
\details{
The miss rate \code{\link{mirt}} and sensitivity \code{\link{sens}}
are complements (\code{mirt = 1 - sens}) and both features of
the decision process (e.g., a diagnostic test).

The function \code{comp_mirt} is complementary to the conversion function
\code{\link{comp_sens}} and uses the generic function
\code{\link{comp_prob_comp}}.
}
\examples{
comp_mirt(2)                      # => NA + warning (beyond range)
comp_mirt(1/3)                    # => 0.6666667
comp_mirt(comp_prob_comp(0.123))  # => 0.123

}
\seealso{
\code{\link{comp_prob_comp}} computes a probability's complement;
\code{\link{is_complement}} verifies probability complements;
\code{\link{comp_prob}} computes current probability information;
\code{\link{prob}} contains current probability information.

Other functions computing probabilities: \code{\link{FDR}},
  \code{\link{FOR}}, \code{\link{NPV}}, \code{\link{PPV}},
  \code{\link{comp_comp_pair}},
  \code{\link{comp_complete_prob_set}},
  \code{\link{comp_fart}}, \code{\link{comp_prob_comp}},
  \code{\link{comp_prob}}, \code{\link{comp_sens}},
  \code{\link{comp_spec}}
}
