% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_prob_freq.R
\name{comp_prob_freq}
\alias{comp_prob_freq}
\title{Compute probabilities from frequencies.}
\usage{
comp_prob_freq(hi = freq$hi, mi = freq$mi, fa = freq$fa, cr = freq$cr)
}
\arguments{
\item{hi}{The number true positives, or hits \code{\link{hi}}}

\item{mi}{The number false negatives, or misses \code{\link{mi}}}

\item{fa}{The number false positives, or false alarms \code{\link{fa}}}

\item{cr}{The number true negatives, or correct rejections \code{\link{cr}}}

\item{N}{The population size \code{\link{N}} (not used yet, use for scaling to new population sizes)}
}
\description{
\code{comp_prob_freq} computes all probabilities
contained in \code{\link{prob}} from 4 essential frequencies
(\code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, \code{\link{cr}}).
}
\details{
\code{comp_prob_freq} is an analog to \code{\link{comp_freq_freq}}
that computes all \emph{frequencies} contained in \code{\link{freq}}
from the same 4 essential frequencies.

The following relationships hold (and are used in computations):

\enumerate{

\item prevalence \code{\link{prev}}:

\code{\link{prev} = \link{cond.true}/\link{N}  =  (\link{hi} + \link{mi}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item sensitivity \code{\link{sens}}:

\code{\link{sens} = \link{hi}/\link{cond.true}  =  \link{hi} / (\link{hi} + \link{mi})  =  (1 - \link{mirt})}

\item miss rate \code{\link{mirt}}:

\code{\link{mirt} = \link{mi}/\link{cond.true}  =  \link{mi} / (\link{hi} + \link{mi})  =  (1 - \link{sens})}

\item specificity \code{\link{spec}}:

\code{\link{spec} = \link{cr}/\link{cond.false}  =  \link{cr} / (\link{fa} + \link{cr})  =  (1 - \link{fart})}

\item false alarm rate \code{\link{fart}}:

\code{\link{fart} = \link{fa}/\link{cond.false}  =  \link{fa} / (\link{fa} + \link{cr})  =  (1 - \link{spec})}



\item proportion of positive decisions \code{\link{ppod}}:

\code{\link{ppod} = \link{dec.pos}/\link{N}  =  (\link{hi} + \link{fa}) / (\link{hi} + \link{mi} + \link{fa} + \link{cr})}

\item positive predictive value \code{\link{PPV}}:

\code{\link{PPV} = \link{hi}/\link{dec.pos}  =  \link{hi} / (\link{hi} + \link{fa})  =  (1 - \link{FDR})}

\item negative predictive value \code{\link{NPV}}:

\code{\link{NPV} = \link{cr}/\link{dec.neg}  =  \link{cr} / (\link{mi} + \link{cr})  =  (1 - \link{FOR})}

\item false detection rate \code{\link{FDR}}:

\code{\link{FDR} = \link{fa}/\link{dec.pos}  =  \link{fa} / (\link{hi} + \link{fa})  =  (1 - \link{PPV})}

\item false omission rate \code{\link{FOR}}:

\code{\link{FOR} = \link{mi}/\link{dec.neg}  =  \link{mi} / (\link{mi} + \link{cr})  =  (1 - \link{NPV})}

}
}
\examples{
## Basics:
comp_prob_freq()
all.equal(prob, comp_prob_freq())  # => should be TRUE!


## Circular chain:
# 1. Current numeric parameters:
num

# 2. Compute all 10 probabilities in prob (from essential probabilities):
prob <- comp_prob()

# 3. Compute 9 frequencies in freq from probabilities:
freq <- comp_freq(round = FALSE)   # prevent rounding (to obtain same probabilities later)
freq

# 4. Compute all 10 probabilities again (but now from frequencies):
prob_freq <- comp_prob_freq()
prob_freq

# 5. Check equality of results (steps 2. and 4.):
all.equal(prob, prob_freq)  # => should be TRUE


}
\seealso{
\code{\link{comp_freq_freq}} computes full frequency information from essential frequencies;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{is_freq}} verifies frequency inputs.

Other functions computing probabilities: \code{\link{comp_FDR}},
  \code{\link{comp_FOR}}, \code{\link{comp_NPV}},
  \code{\link{comp_PPV}}, \code{\link{comp_comp_pair}},
  \code{\link{comp_complete_prob_set}},
  \code{\link{comp_fart}}, \code{\link{comp_mirt}},
  \code{\link{comp_prob_comp}}, \code{\link{comp_prob}},
  \code{\link{comp_sens}}, \code{\link{comp_spec}}
}
