% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_prob_prob.R
\name{comp_sens}
\alias{comp_sens}
\title{Compute a decision's sensitivity from its miss rate.}
\usage{
comp_sens(mirt)
}
\arguments{
\item{mirt}{The decision's miss rate \code{\link{mirt}}.}
}
\value{
The decision's sensitivity \code{\link{sens}}.
}
\description{
\code{comp_sens} is a conversion function that takes a miss rate \code{\link{mirt}}
-- given as a probability (i.e., a numeric value in the range from 0 to 1) --
as its input, and returns the corresponding sensitivity \code{\link{sens}}
-- also as a probability -- as its output.
}
\details{
The sensitivity \code{\link{sens}} and miss rate \code{\link{mirt}}
are complements (\code{sens = 1 - mirt}) and both features of
the decision process (e.g., a diagnostic test).

The function \code{comp_sens} is complementary to the conversion function
\code{\link{comp_mirt}} and uses the generic function
\code{\link{comp_prob_comp}}.
}
\examples{
comp_sens(2)                      # => NA + warning (beyond range)
comp_sens(1/3)                    # => 0.6666667
comp_sens(comp_prob_comp(0.123))  # => 0.123

}
\seealso{
\code{\link{comp_prob_comp}} computes a probability's complement;
\code{\link{is_complement}} verifies probability complements;
\code{\link{comp_prob}} computes current probability information;
\code{\link{prob}} contains current probability information.

Other functions computing probabilities: \code{\link{FDR}},
  \code{\link{FOR}}, \code{\link{NPV}}, \code{\link{PPV}},
  \code{\link{comp_comp_pair}}, \code{\link{comp_fart}},
  \code{\link{comp_mirt}}, \code{\link{comp_prob_comp}},
  \code{\link{comp_prob}}, \code{\link{comp_spec}}
}
