% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{cr}
\alias{cr}
\alias{TN}
\title{Frequency of correct rejections or true negatives (TN).}
\format{An object of class \code{logical} of length 1.}
\usage{
cr
}
\description{
\code{cr} is the frequency of correct rejections
or true negatives (\code{TN})
in a population of \code{\link{N}} individuals.
}
\details{
Definition: \code{cr} are individuals for which
\code{cond.false} and \code{dec.neg}
holds simultaneously.

\code{cr} are cases of correct classifications.

Relationships:
In a population of size \code{\link{N}}
the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}}

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}}

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}}
}
}
\seealso{
\code{\link{spec}} is the specificity or correct rejection rate
(aka. true negative rate \code{\link{TNR}});
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{is_freq}} verifies frequencies.

Other essential parameters: \code{\link{fa}},
  \code{\link{hi}}, \code{\link{mi}}

Other frequencies: \code{\link{N}}, \code{\link{fa}},
  \code{\link{hi}}, \code{\link{mi}}
}
\keyword{datasets}
