% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{dec_cor}
\alias{dec_cor}
\title{Number of individuals for which the decision is correct.}
\format{An object of class \code{numeric} of length 1.}
\usage{
dec_cor
}
\description{
\code{dec_cor} is a frequency that describes the
number of individuals in the current population \code{\link{N}}
for which the decision is correct/accurate
(i.e., cases in which the decision corresponds to the condition).
}
\details{
Key relationships:

\enumerate{

\item to probabilities:
The frequency of \code{dec_cor} individuals depends on the population size \code{\link{N}} and
the accuracy \code{\link{acc}}.

\item to other frequencies:
In a population of size \code{\link{N}} the following relationships hold:

  \itemize{

    \item \code{\link{N}  =  \link{cond_true} + \link{cond_false}} (by condition)

    \item \code{\link{N}  =  \link{dec_pos} + \link{dec_neg}} (by decision)

    \item \code{\link{N} = \link{dec_cor} + \link{dec_err}} (by correspondence of decision to condition)

    \item \code{\link{dec_cor} = \link{hi} + \link{cr}}

    \item \code{\link{dec_err} = \link{mi} + \link{fa}}

    \item \code{\link{N}  =  \link{hi} + \link{mi} + \link{fa} + \link{cr}} (by condition x decision)
  }

\item correspondence:
When not rounding the frequencies of \code{\link{freq}} then

\code{dec_cor = N x acc = hi + cr}

(i.e., \code{dec_cor} corresponds to the sum of true positives \code{\link{hi}} and true negatives \code{\link{cr}}.

}

Current frequency information is computed by
\code{\link{comp_freq}} and contained in a list
\code{\link{freq}}.
}
\examples{
dec_cor <- 1000 * .50   # => sets dec_cor to 50\% of 1000 = 500 cases.
is_freq(dec_cor)        # => TRUE
is_prob(dec_cor)        # => FALSE, as dec_cor is no probability (but acc, bacc/wacc ARE)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Confusion_matrix}{Wikipedia: Confusion matrix} for additional information.
}
\seealso{
\code{\link{is_freq}} verifies frequencies;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information.

Other frequencies: 
\code{\link{N}},
\code{\link{cond_false}},
\code{\link{cond_true}},
\code{\link{cr}},
\code{\link{dec_err}},
\code{\link{dec_neg}},
\code{\link{dec_pos}},
\code{\link{fa}},
\code{\link{hi}},
\code{\link{mi}}
}
\concept{frequencies}
\keyword{datasets}
