% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A}
\alias{df_A}
\title{Cumulative risk curve A (main/transfer task A).}
\format{
A data frame (17 x 2).

\code{age}: age (in years).

\code{crisk_A}: cumulative risk of developing
some disease A in the target population.
}
\usage{
df_A
}
\description{
\code{df_A} provides the cumulative risk of
some genetic risk factor for developing disease A
in some target population as a function of age.
}
\seealso{
\code{\link{plot_crisk}} plots cumulative risk curves.

Other datasets: 
\code{\link{BRCA1}},
\code{\link{BRCA2}},
\code{\link{df_B}},
\code{\link{df_I}},
\code{\link{df_scenarios}}
}
\concept{datasets}
\keyword{datasets}
