% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_B}
\alias{df_B}
\title{Cumulative risk curve B (main/transfer task B).}
\format{
A data frame (17 x 2).

\code{age}: age (in years).

\code{crisk_B}: cumulative risk of developing
some disease B in the target population.
}
\usage{
df_B
}
\description{
\code{df_B} provides the cumulative risk of
some genetic risk factor for developing disease B
in some target population as a function of age.
}
\seealso{
\code{\link{plot_crisk}} plots cumulative risk curves.

Other datasets: 
\code{\link{BRCA1}},
\code{\link{BRCA2}},
\code{\link{df_A}},
\code{\link{df_I}},
\code{\link{df_scenarios}}
}
\concept{datasets}
\keyword{datasets}
