% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_I}
\alias{df_I}
\title{Cumulative risk curve I (introductory task).}
\format{
A data frame (17 x 2).

\code{age}: age (in years).

\code{crisk_I}: cumulative risk of developing
some disease in the target population.
}
\usage{
df_I
}
\description{
\code{df_I} provides the cumulative risk of
some genetic risk factor for developing a disease
in some target population as a function of age.
}
\seealso{
\code{\link{plot_crisk}} plots cumulative risk curves.

Other datasets: 
\code{\link{BRCA1}},
\code{\link{BRCA2}},
\code{\link{df_A}},
\code{\link{df_B}},
\code{\link{df_scenarios}}
}
\concept{datasets}
\keyword{datasets}
