% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{fa}
\alias{fa}
\alias{FP}
\title{Frequency of false alarms or false positives (FP).}
\format{An object of class \code{logical} of length 1.}
\usage{
fa
}
\description{
\code{fa} is the frequency of false alarms
or false positives (\code{FP})
in a population of \code{\link{N}} individuals.
}
\details{
Definition: \code{fa} are individuals for which
\code{cond.false} and \code{dec.pos}
holds simultaneously.

\code{fa} are cases of incorrect classifications
(type-I-errors).

Relationships:
In a population of size \code{\link{N}}
the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}}

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}}

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}}
}
}
\seealso{
\code{\link{fart}} is the probability of false alarms
(aka. false positive rate \code{\link{FPR}} or \code{\link{fallout}});
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{is_freq}} verifies frequencies.

Other essential parameters: \code{\link{cr}},
  \code{\link{hi}}, \code{\link{mi}}

Other frequencies: \code{\link{N}}, \code{\link{cr}},
  \code{\link{hi}}, \code{\link{mi}}
}
\keyword{datasets}
