% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prob.R
\docType{data}
\name{FDR}
\alias{FDR}
\title{The false detection rate of a decision process or diagnostic procedure.}
\format{An object of class \code{logical} of length 1.}
\usage{
FDR
}
\description{
\code{FDR} defines a decision's false detection (or false discovery)
rate (\code{FDR}): The conditional probability of the condition
being \code{FALSE} provided that the decision is positive.
}
\details{
Understanding or obtaining the false detection fate
or false discovery rate (\code{FDR}):

\itemize{

\item Definition:
  \code{FDR} is the conditional probability
  for the condition being \code{FALSE}
  given a positive decision:

\code{FDR = p(condition = FALSE | decision = positive)}

\item Relationships:

a. \code{FDR} is the complement of the
  positive predictive value \code{\link{PPV}}:

\code{FDR = 1 - PPV}

b. \code{FDR} is the opposite conditional probability
  -- but not the complement --
  of the false alarm rate \code{\link{fart}}:

\code{fart = p(decision = positive | condition = FALSE)}

\item In natural frequencies,
  \code{FDR} is the ratio of individuals for which
  \code{decision = positive} and \code{condition = FALSE}
  (aka. false positives)
  divided by the number of all individuals for which
  \code{decision = positive} (true positives and false positives):

\code{FDR = n(decision = positive & condition = FALSE) / n(decision = positive)}

\code{FDR = n.fa / (n.hi + n.fa) }

\item \code{FDR} is a feature of a decision process
  or diagnostic procedure and
  a measure of incorrect decisions (positive decisions
  that are actually \code{FALSE}).

However, due to being a conditional probability,
  the value of \code{FDR} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
FDR <- .45     # => sets a false discovery rate (FDR) of 45\%
FDR <- 45/100  # => (condition = FALSE) for 45 people out of 100 people for which (decision = positive)
is_prob(FDR)   # => TRUE (as FDR is a probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values}{Wikipedia} for additional information.
}
\seealso{
\code{\link{comp_FDR}} computes \code{FDR} as the complement of \code{\link{PPV}};
\code{\link{num}} contains basic numeric variables;
\code{\link{comp_PPV}} and \code{\link{comp_PPV_freq}} compute PPVs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities.

Other functions computing probabilities: \code{\link{FOR}},
  \code{\link{NPV}}, \code{\link{PPV}},
  \code{\link{comp_comp_pair}}, \code{\link{comp_fart}},
  \code{\link{comp_mirt}}, \code{\link{comp_prob_comp}},
  \code{\link{comp_prob}}, \code{\link{comp_sens}},
  \code{\link{comp_spec}}
}
\keyword{datasets}
