% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mlm.R
\name{focus}
\alias{focus}
\title{Focus on a 2x2 matrix to compute some measure(s).}
\usage{
focus(mx, measures = c("all"), as_pc = FALSE, n_digits = 3)
}
\arguments{
\item{mx}{A 2x2 matrix (as numeric contingency table, required).}

\item{measures}{A vector of the measure(s) to compute.
Default: \code{measures = c("all")} (i.e., all available measures).}

\item{as_pc}{Boolean: Convert probabilities into percentages?
Default: \code{as_pc = FALSE}.}

\item{n_digits}{Number of decimal places to which result is rounded.
Default: \code{n_digits = 3}.}
}
\value{
A named numeric vector.
}
\description{
\code{focus} adopts particular perspectives on a 2x2 matrix \code{mx}
to compute various scientific measures.
}
\details{
If no subset of \code{measures} is specified,
\code{focus} computes the 36 measures for a 2x2 matrix \code{mx}
that are contained in Table 3 of Neth et al., (2021)
(\url{https://doi.org/10.3389/fpsyg.2020.567817}).
}
\examples{
# 1. The mammography problem:
abcd <- c(8, 95, 2, 895)  # Frequencies (Gigerenzer & Hoffrage, 1995)
mp <- frame(data = abcd, x = "Condition", y = "Test",
            x_levels = c("cancer", "no cancer"),
            y_levels = c("positive", "negative"))

# 1. Computing ALL measures at once (and selecting afterwards):
focus(mp)
focus(mp)["PPV"]
focus(mp)[c("dPc", "dPr", "Chi")]

# 2. Computing individual measures:
# 4 frequencies:
focus(mp, measures = c("TP", "fp", "fn", "TN"))
sum(focus(mp, measures = c("TP", "fp", "fn", "TN")))  # N

# 3 marginal probabilities:
focus(mp, measures = c("prev", "bias", "Acc"))

# 8 conditional probabilities:
focus(mp, measures = c("sens", "fpr", "fnr", "spec"))  # X/by-col
trans(mp, margin = 2)
focus(mp, measures = c("PPV", "FDR", "FOR", "NPV"))    # Y/by-row
trans(mp, margin = 1)

# 3 triangular measures:
focus(mp, measures = c("Jaccard", "F1", "G2"))
# 2 mixed scores:
focus(mp, measures = c("lift", "RI"))

# 8 difference-based measures:
focus(mp, measures = c("dPc", "NNT", "BACC", "RRR",
                       "dPr", "kappa", "MCC", "Chi"))

# 3 simple odds:
focus(mp, measures = c("pre-test odds", "post-test odds+", "post-test odds-"))

# 5 odds ratios:
focus(mp, measures = c("LR+", "LR-", "DOR", "Q", "Y"))

# Typical uses:
focus(mp, measures = c("sens", "spec", "ppv", "npv"))

}
\seealso{
\code{\link{frame}} creates a 2x2 matrix;
\code{\link{trans}} converts a 2x2 matrix (into a table of probabilities/conditional probabilities);
\code{\link{is_matrix}} verifies a 2x2 matrix.

Other matrix lens model functions: 
\code{\link{frame}()},
\code{\link{trans}()}
}
\concept{matrix lens model functions}
