% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_prob.R
\docType{data}
\name{FOR}
\alias{FOR}
\title{The false omission rate (FOR) of a decision process or diagnostic procedure.}
\format{An object of class \code{logical} of length 1.}
\usage{
FOR
}
\description{
\code{FOR} defines a decision's false omission rate (\code{FOR}):
The conditional probability of the condition being \code{TRUE}
provided that the decision is negative.
}
\details{
Understanding or obtaining the false omission rate \code{FOR}:

\itemize{

\item Definition:
  \code{FOR} is the conditional probability
  for the condition being \code{TRUE}
  given a negative decision:

\code{FOR = p(condition = TRUE | decision = negative)}

\item Relationships:

a. \code{FOR} is the complement of the
  negative predictive value \code{\link{NPV}}:

\code{FOR = 1 - NPV}

b. \code{FOR} is the opposite conditional probability
  -- but not the complement --
  of the missing rate (\code{1 - \link{sens}}):

\code{missing rate = p(decision = negative | condition = TRUE)}

\item In natural frequencies,
  \code{FOR} is the ratio of individuals for which
  \code{decision = negative} and \code{condition = TRUE}
  (aka. false negatives)
  divided by the number of all individuals for which
  \code{decision = negative} (true negatives and false negatives):

\code{FOR = n(decision = negative & condition = TRUE) / n(decision = negative)}

\code{FOR = n.mi / (n.mi + n.cr)}

\item \code{FOR} is a feature of a decision process
  or diagnostic procedure and a measure of incorrect
  decisions (negative decisions that are actually \code{FALSE}).

However, due to being a conditional probability,
  the value of \code{FOR} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
FOR <- .05     # => sets a false omission rate of 5\%
FOR <- 5/100   # => (condition = TRUE) for 5 people out of 100 people for which (decision = negative)
is_prob(FOR)   # => TRUE (as FOR is a probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values}{Wikipedia} for additional information.
}
\seealso{
\code{\link{comp_FOR}} computes \code{FOR} as the complement of \code{\link{NPV}};
\code{\link{num}} contains basic numeric variables;
\code{\link{comp_NPV}} and \code{\link{comp_NPV_freq}} compute PPVs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities

Other functions computing probabilities: \code{\link{FDR}},
  \code{\link{NPV}}, \code{\link{PPV}},
  \code{\link{comp_complement}}, \code{\link{comp_fart}},
  \code{\link{comp_prob}}, \code{\link{comp_spec}}
}
\keyword{datasets}
