% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mlm.R
\name{frame}
\alias{frame}
\title{Frame a 2x2 matrix (from data or description).}
\usage{
frame(
  data,
  x,
  y,
  z = NA,
  z_val = NA,
  freq_var = NA,
  x_name = NA,
  y_name = NA,
  x_levels = NA,
  y_levels = NA
)
}
\arguments{
\item{data}{A data frame (with at least 2 binary variables \code{x} and \code{y},
denoting individual cases or a contingency table with frequency counts)
or a vector with 4 basic frequency counts (abcd, in by-row direction).}

\item{x}{Variable of \code{data} used for x-dimension of 2x2 matrix.}

\item{y}{Variable of \code{data} used for y-dimension of 2x2 matrix.}

\item{z}{Variable of \code{data} used for conditionalizing \code{data}.}

\item{z_val}{Value of \code{z} used for conditionalizing \code{data}.}

\item{freq_var}{Variable containing the frequency count of each case
when \code{data} provides a contingency table
(i.e., each row provides a case description).}

\item{x_name}{Name of x-dimension in 2x2 matrix.}

\item{y_name}{Name of y-dimension in 2x2 matrix.}

\item{x_levels}{Order of category levels on x-dimension
(values must be present in \code{data}).}

\item{y_levels}{Order of category levels on y-dimension
(values must be present in \code{data}).}
}
\value{
A 2x2 matrix (as a contingency \code{table} of integer values).
}
\description{
\code{frame} creates a 2x2 matrix from data or
a description of its contents and layout.
}
\details{
\code{frame} supports 3 use cases:

\enumerate{

\item From raw data (as data frame):
Each row of \code{data} denotes an individual case,
and the columns contain at least 2 binary variables \code{x} and \code{y}.

\item From a contingency table (as data frame):
Each row of \code{data} denotes a case description,
a numeric variable \code{freq_var} denotes the frequency count per case,
and other columns contain at least 2 binary variables \code{x} and \code{y}.

\item From a description:
A numeric vector \code{data} provides 4 basic frequency counts
(i.e., abcd, read in by-row direction),
and a specification of the matrix layout
(i.e., names of dimensions \code{x} and \code{y},
and the names/orders of category levels).

}

Use cases 1. and 2. allow conditionalizing \code{data}
on a value \code{z_val} of a variable \code{z}.
}
\examples{
# 1. From raw data (denoting individual cases):
s <- riskyr(hi = 8, mi = 2, fa = 95, cr = 895)
d <- write_popu(s)
frame(d, x = "True condition", y = "Outcome")

# 2. From a contingency table (df) with a variable denoting
#    the frequency counts of each case:
c_tab <- as.data.frame(Titanic)
frame(c_tab, x = "Sex", y = "Survived", freq_var = "Freq")
frame(c_tab, x = "Sex", y = "Survived", freq_var = "Freq",
      x_name = "Gender", y_name = "Alive",
      x_levels = c("Female", "Male"), y_levels = c("Yes", "No"))

# 3. From a description of 4 frequency counts and matrix layout:
# a. Basics:
frame(data = 1:4, x = "Condition", y = "Outcome")

# b. Mammography problem:
abcd <- c(8, 95, 2, 895)  # 4 frequencies (Gigerenzer & Hoffrage, 1995)
frame(data = abcd, x = "Condition", y = "Test",
      x_levels = c("cancer", "no cancer"),
      y_levels = c("positive", "negative"))

# c. Titanic passengers:
abcd <- c(344, 367, 126, 1364)
frame(data = abcd, x = "sex", y = "survived",
      x_levels = c("female", "male"), y_levels = c(1, 0))
frame(data = abcd, x = "Gender", y = "Alive",
      x_levels = c("Female", "Male"), y_levels = c("Yes", "No"))

}
\seealso{
\code{\link{comp_popu}} creates data (as df) from description (frequencies);
\code{\link{read_popu}} creates a scenario (description) from data (as df);
\code{\link{riskyr}} initializes a \code{riskyr} scenario.
}
\concept{matrix lens model functions}
