% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num_freq.R
\docType{data}
\name{freq}
\alias{freq}
\title{List current frequency information.}
\format{An object of class \code{list} of length 9.}
\usage{
freq
}
\description{
\code{freq} is a list of named numeric variables
containing 9 (natural) frequencies:
\enumerate{
 \item the population size \code{\link{N}}
 \item the number of cases for which \code{cond.true}
 \item the number of cases for which \code{cond.false}
 \item the number of cases for which \code{dec.pos}
 \item the number of cases for which \code{dec.neg}
 \item the number true positives, or hits (\code{hi})
 \item the number false negatives, or misses (\code{mi})
 \item the number false positives, or false alarms (\code{fa})
 \item the number true negatives, or correct rejections (\code{cr})
}
}
\details{
These frequencies are computed from basic probabilities
(contained in \code{\link{num}}) and computed by using
\code{\link{comp_freq}}.

The list \code{freq} is the frequency counterpart
to the list containing probability information \code{\link{prob}}.

Natural frequencies are always expressed in
relation to the current population of
size \code{\link{N}}.

Key relationships:

\enumerate{

\item to probabilities:
A population of \code{\link{N}} individuals can be split into 2 subsets
in 2 different ways:

\enumerate{
  \item by condition:
  The frequency \code{\link{cond.true}} depends on the prevalence \code{\link{prev}}
  and
  the frequency \code{\link{cond.false}} depends on the prevalence's complement \code{1 - \link{prev}}.

\item by decision:
  The frequency \code{\link{dec.pos}} depends on the proportion of positive decisions \code{\link{ppod}}
  and
  the frequency \code{\link{dec.neg}} depends on the proportion of negative decisions \code{1 - \link{ppod}}.

}

The population size \code{\link{N}} is a free parameter (independent of the
essential probabilities \code{\link{prev}}, \code{\link{sens}}, and \code{\link{spec}}).

If \code{\link{N}} is unknown, a suitable minimum value can be computed by \code{\link{comp_min_N}}.

\item to other frequencies:
  In a population of size \code{\link{N}}
  the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}} (by condition)

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}} (by decision)

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}} (by condition x decision)
  }

}

Visualizations of the current frequency information
are provided by \code{\link{plot_tree}} and
\code{\link{plot_mosaic}}.
}
\examples{
freq <- comp_freq()  # => initialize freq to default parameters
freq                 # => show current values
length(freq)         # => 9

}
\seealso{
\code{\link{comp_freq}} computes current frequency information;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{prob}} contains current probability information

Other lists containing scenario settings: \code{\link{num}},
  \code{\link{prob}}
}
\keyword{datasets}
