% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{hi}
\alias{hi}
\alias{TP}
\title{Frequency of hits or true positives (TP).}
\format{An object of class \code{logical} of length 1.}
\usage{
hi
}
\description{
\code{hi} is the frequency of hits or true positives (TP)
in a population of \code{\link{N}} individuals.
}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{freq}} contains basic frequency variables.

Other basic frequencies: \code{\link{N}}
}
\keyword{datasets}
