% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\name{init_num}
\alias{init_num}
\title{Initialize basic numeric variables.}
\usage{
init_num(prev = num.def$prev, sens = num.def$sens, spec = num.def$spec,
  fart = num.def$fart, N = num.def$N)
}
\arguments{
\item{prev}{The condition's prevalence value (i.e., the probability of condition being TRUE).}

\item{sens}{A decision's sensitivity value (i.e., the conditional probability
of a positive decision provided that the condition is TRUE).}

\item{spec}{A specificity value (i.e., the conditional probability
of a negative decision provided that the condition is FALSE).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{A false alarm rate (i.e., the conditional probability
of a positive decision provided that the condition is FALSE).
\code{fart} is optional when its complement \code{spec} is provided.}
}
\value{
A list containing a valid quadruple of probabilities
(\code{prev}, \code{sens}, \code{spec}, and \code{fart})
and one frequency (population size \code{N}).
}
\description{
\code{init_num} initializes basic numeric variables to define \code{num}
as a list of named elements containing four basic probabilities
(\code{prev}, \code{sens}, \code{spec}, and \code{fart})
and one frequency parameter (the population size \code{N}).
}
\details{
If \code{spec} is provided, its complement \code{fart} is optional.
If \code{fart} is provided, its complement \code{spec} is optional.
If no \code{N} is provided, it is computed by \code{\link{comp_min_N}}.
}
\examples{
init_num(1, 1, 1, 0, 100)  # => succeeds
init_num(1, 1, 0, 1, 100)  # => succeeds

init_num(1, 1, 0, 1)           # => succeeds (with N computed)
init_num(1, 1, NA, 1, 100)     # => succeeds (with spec computed)
init_num(1, 1, 0, NA, 100)     # => succeeds (with fart computed)
init_num(1, 1, NA, 1)          # => succeeds (with spec and N computed)
init_num(1, 1, 0, NA)          # => succeeds (with fart and N computed)
init_num(1, 1, .51, .50, 100)  # => succeeds (as spec and fart are within tolarated range)

init_num(prev = NA)                   # => NAs + warning that prev is not numeric
init_num(prev = 88)                   # => NAs + warning that prev is no probability
init_num(prev =  1, sens = NA)        # => NAs + warning that sens is not numeric
init_num(1, 1, spec = NA, fart = NA)  # => NAs + warning that spec or fart is necessary
init_num(1, 1, 1, 1)                  # => NAs + warning that spec and fart are not complements (in tolerated range)
init_num(1, 1, .52, .50, 100)         # => NAs + warning that spec and fart are not complements (in tolerated range)

}
\seealso{
\code{\link{num}} to store basic parameter values;
\code{\link{comp_min_N}} to get a minimum value of population size N

Other functions to initialize scenario settings: \code{\link{init_pal}},
  \code{\link{init_txt}}
}
