% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_complement}
\alias{is_complement}
\title{Verify that two arguments are numeric complements.}
\usage{
is_complement(spec, fart, tol = 0.01)
}
\arguments{
\item{spec}{A specificity value (i.e., the conditional probability
of a negative decision provided that the condition is FALSE).}

\item{fart}{A false alarm rate (i.e., the conditional probability
of a positive decision provided that the condition is FALSE).}

\item{tol}{A numeric tolerance value.}
}
\value{
A Boolean value: \code{TRUE} if the arguments are complements,
otherwise \code{FALSE}.
}
\description{
\code{is_complement} is a function that
takes 2 numeric arguments (probabilities) as inputs and
verifies that they are complements (i.e., add up to 1).
}
\details{
Both \code{spec} and \code{fart} are necessary arguments.

The argument \code{tol} is optional (with a default value of .01)
Complements differing by less than this
value are still considered to be complements.

This function does not verify the type, range or sufficiency
of the inputs provided. See \code{\link{is_prob}} and
\code{\link{is_sufficient}} for this purpose.
}
\examples{
is_complement(0, 0)              # => FALSE + Warning that difference exceeds tolerance

is_complement(0, 1)              # => TRUE
is_complement(1/3, 2/3)          # => TRUE
is_complement(.33, .66)          # => TRUE

is_complement(.3, .6)            # => FALSE + Warning that difference exceeds tolerance
is_complement(.3, .6, tol = .1)  # => TRUE (due to increased tolerance)

}
\seealso{
\code{\link{as_pc}} to display a probability as a percentage

Other verification functions: \code{\link{is_perc}},
  \code{\link{is_prob}}, \code{\link{is_sufficient}},
  \code{\link{is_valid}}
}
