% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_extreme}
\alias{is_extreme}
\title{Verify that probabilities provided describe an extreme case.}
\usage{
is_extreme(prev, sens, spec = NA, fart = NA)
}
\arguments{
\item{prev}{The condition's prevalence value \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity value \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{FALSE}).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}
}
\value{
A Boolean value: \code{TRUE} if an extreme case is identified,
otherwise \code{FALSE}.
}
\description{
\code{is_extreme} verifies that the current combination
of probabilities provided (i.e., \code{\link{prev}}
and \code{\link{sens}}, and \code{\link{spec}} or
\code{\link{fart}}) describe an extreme case.
}
\details{
If \code{TRUE}, a warning message describing the
nature of the extreme case is printed to allow
anticipating peculiar effects (e.g., that
PPV or NPV values cannot be computed or are \code{NaN}).

This function does not verify the type, range, sufficiency,
or consistency of its arguments. See \code{\link{is_prob}},
\code{\link{is_sufficient}}, and \code{\link{is_complement}}
for this purpose.
}
\examples{
# Identify 4 extreme cases (+2 variants):
is_extreme(1, 1)           # => TRUE + warning: N true positives
plot_tree(1, 1, N = 100)   # => illustrates this case

is_extreme(1, 0)           # => TRUE + warning: N false negatives
plot_tree(1, 0, N = 100)   # => illustrates this case

sens <- .50
is_extreme(0, sens, 0)              # => TRUE + warning: N false positives
plot_tree(0, sens, 0, N = 100)      # => illustrates this case
# Variant:
is_extreme(0, sens, NA, 1)          # => TRUE + warning: N false positives
plot_tree(0, sens, NA, 1, N = 100)  # => illustrates this case

is_extreme(0, sens, 1)              # => TRUE + warning: N true negatives
plot_tree(0, sens, 1, N = 100)      # => illustrates this case
# Variant:
is_extreme(0, sens, NA, 0)          # => TRUE + warning: N true negatives
plot_tree(0, sens, NA, 0, N = 100)  # => illustrates this case

}
\seealso{
\code{\link{is_valid}} uses \code{is_extreme}
to verify the validity of probability inputs;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{as_pc}} displays a probability as a percentage;
\code{\link{as_pb}} displays a percentage as probability

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_freq}}, \code{\link{is_perc}},
  \code{\link{is_prob}}, \code{\link{is_sufficient}},
  \code{\link{is_valid}}
}
