% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_perc}
\alias{is_perc}
\title{Verify that a numeric input is a percentage (from 0 to 100).}
\usage{
is_perc(perc)
}
\arguments{
\item{perc}{A single (typically numeric) argument.}
}
\value{
A Boolean value: \code{TRUE} if \code{perc} is a percentage (proportion),
otherwise \code{FALSE}.
}
\description{
\code{is_perc} is a function that checks whether its single argument \code{perc}
is a percentage (proportion, i.e., a numeric value in the range from 0 to 100).
}
\examples{
is_perc("Bernoulli") # => FALSE + warning that perc is not numeric

is_perc(2)           # => TRUE, but does NOT return the percentage 2.
is_perc(1/2)         # => TRUE, but does NOT return the percentage 0.5.

perc.seq <- seq(0, 100, by = 10)
is_perc(perc.seq)    # => TRUE (for vector)
}
\seealso{
\code{\link{as_pc}} to display a probability as a percentage;
\code{\link{as_pb}} to display a percentage as probability

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_prob}}, \code{\link{is_sufficient}},
  \code{\link{is_valid}}
}
