% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_perfect}
\alias{is_perfect}
\title{Verify that a condition's prevalence and a decision's
sensitivity are both perfect.}
\usage{
is_perfect(prev, sens)
}
\arguments{
\item{prev}{The condition's prevalence value \code{\link{prev}}
(i.e., the probability of condition being TRUE).}

\item{sens}{The decision's sensitivity value \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is TRUE).}
}
\value{
A Boolean value: \code{TRUE} if \code{\link{prev}}
and \code{\link{sens}} are both 1, otherwise \code{FALSE}.
}
\description{
\code{is_perfect} verifies that the current prevalence
value \code{\link{prev}} and sensitivity value \code{\link{sens}}
are both perfect (i.e., both have a maximum value of 1)
at the same time.
}
\details{
If \code{TRUE}, a warning message is printed,
as there are no false decisions (i.e., no false negatives
and no false positives) and some derived probabilities
(e.g., the conditional probabilities \code{\link{npv}}
and \code{\link{FOR}}) cannot be computed.
}
\examples{
is_perfect(1, 1)           # => TRUE + warning (as npv cannot be computed)

is_perfect(1, 2)           # => FALSE
is_perfect(1, 1 - 10^-12)  # => FALSE

}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{is_valid}} verifies the validity of probability inputs;
\code{\link{as_pc}} displays a probability as a percentage;
\code{\link{as_pb}} displays a percentage as probability

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_freq}}, \code{\link{is_perc}},
  \code{\link{is_prob}}, \code{\link{is_sufficient}},
  \code{\link{is_valid}}
}
