% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_prob}
\alias{is_prob}
\title{Verify that a numeric input is a probability (from 0 to 1).}
\usage{
is_prob(prob)
}
\arguments{
\item{prob}{A numeric argument (scalar or vector).}
}
\value{
A Boolean value: \code{TRUE} if \code{prob} is a probability, otherwise \code{FALSE}.
}
\description{
\code{is_prob} is a function that checks whether its argument \code{prob}
is a probability (i.e., a numeric value in the range from 0 to 1).
}
\examples{
is_prob("Laplace")  # => FALSE + warning that prob is not numeric
is_prob(2)          # => FALSE + warning that prob is not in 0 to 1 range

is_prob(1/2)        # => TRUE, but does NOT return the probability .5

prob.seq <- seq(0, 1, by = 1/10)
is_prob(prob.seq)   # => TRUE (for vector)

}
\seealso{
\code{\link{as_pc}} to display a probability as a percentage;
\code{\link{as_pb}} to display a percentage as probability

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_perc}}, \code{\link{is_sufficient}},
  \code{\link{is_valid}}
}
