% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_sufficient}
\alias{is_sufficient}
\title{Verify that a sufficient number of probabilities are provided.}
\usage{
is_sufficient(prev, sens, spec = NA, fart = NA)
}
\arguments{
\item{prev}{The condition's prevalence value (i.e., the probability of condition being TRUE).}

\item{sens}{A decision's sensitivity value (i.e., the conditional probability
of a positive decision provided that the condition is TRUE).}

\item{spec}{A specificity value (i.e., the conditional probability
of a negative decision provided that the condition is FALSE).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{A false alarm rate (i.e., the conditional probability
of a positive decision provided that the condition is FALSE).
\code{fart} is optional when its complement \code{spec} is provided.}
}
\value{
A Boolean value: \code{TRUE} if the parameters provided are sufficient,
otherwise \code{FALSE}.
}
\description{
\code{is_sufficient} is a function that
takes 3 or 4 arguments (typically probabilities) as inputs and
verifies that they are sufficient to compute the
frequencies and conditional probabilities
for a population of N individuals.
}
\details{
While no alternative input option for frequencies is provided,
specification of \code{prev} and \code{sens} are always
necessary. One additional probability parameter is necessary:
If \code{spec} is provided, its complement \code{fart} is optional.
If \code{fart} is provided, its complement \code{spec} is optional.

Note that this function does not verify the type, range or
consistency of its arguments. See \code{\link{is_prob}} and
\code{\link{is_complement}} for this purpose.
}
\examples{
is_sufficient()          # => FALSE + Warning
is_sufficient(prev = 1)  # => FALSE + Warning
is_sufficient(prev = 1, sens = 2)  # => FALSE + Warning

is_sufficient(prev = 1, sens = 2, spec = 3)  # => TRUE, but is_prob would be FALSE for 2 and 3
is_sufficient(prev = 1, sens = 2, fart = 4)  # => TRUE, but is_prob would be FALSE for 2 and 4

is_sufficient(prev = 1, sens = 2, spec = 3, fart = 4)  # => TRUE, but is_prob would be FALSE for 2, 3, and 4

}
\seealso{
\code{\link{as_pc}} to display a probability as a percentage

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_perc}}, \code{\link{is_prob}},
  \code{\link{is_valid}}
}
