% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{is_valid}
\alias{is_valid}
\title{Verify that basic probabilities are valid inputs.}
\usage{
is_valid(prev, sens, spec = NA, fart = NA, tol = 0.01)
}
\arguments{
\item{prev}{The condition's prevalence value (i.e., the probability of condition being TRUE).}

\item{sens}{A decision's sensitivity value (i.e., the conditional probability
of a positive decision provided that the condition is TRUE).}

\item{spec}{A specificity value (i.e., the conditional probability
of a negative decision provided that the condition is FALSE).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{A false alarm rate (i.e., the conditional probability
of a positive decision provided that the condition is FALSE).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{tol}{A numeric tolerance value used by \code{\link{is_complement}}.}
}
\value{
A Boolean value: \code{TRUE} if the parameters provided are valid,
otherwise \code{FALSE}.
}
\description{
\code{is_valid} is a function that takes
3 or 4 numeric arguments as inputs and
verifies that they can be interpreted as a
valid quadruple of probabilities.
}
\details{
\code{is_valid} is a wrapper function that combines
\code{\link{is_prob}}, \code{\link{is_sufficient}},
and \code{\link{is_complement}} in one function.

Note that \code{is_valid} only verifies the validity of inputs,
but does not compute or return numeric variables.
Use \code{\link{init_num}} for initializing basic parameters.

Both \code{prev} and \code{sens} and
either \code{spec} or \code{fart} are necessary arguments.
The argument \code{tol} is optional (with a default value of .01)
and used as the tolerance value of \code{\link{is_complement}}.
}
\examples{
is_valid(1, 1, 1, 0)   # => TRUE
is_valid(1, 1, 0, 1)   # => TRUE
is_valid(1, 1, 1, NA)  # => TRUE
is_valid(1, 1, NA, 1)  # => TRUE

is_valid(1, 1, 1, 1)   # => FALSE + warning that is_complement fails
is_valid(1, 1, NA, NA) # => FALSE + warning that is_sufficient fails
is_valid(8, 1, 1, 0)   # => FALSE + warning that is_prob(prev) fails
is_valid(1, 8, 1, 0)   # => FALSE + warning that is_prob(sens) fails
is_valid(1, 1, 8, NA)  # => FALSE + warning that is_prob(spec) fails
is_valid(1, 1, NA, 8)  # => FALSE + warning that is_prob(fart) fails

}
\seealso{
\code{\link{init_num}} for initializing basic parameters

Other verification functions: \code{\link{is_complement}},
  \code{\link{is_perc}}, \code{\link{is_prob}},
  \code{\link{is_sufficient}}
}
