% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{mi}
\alias{mi}
\alias{FN}
\alias{type-II-errors}
\title{Frequency of misses or false negatives (FN).}
\format{An object of class \code{logical} of length 1.}
\usage{
mi
}
\description{
\code{mi} is the frequency of misses
or false negatives (\code{FN})
in a population of \code{\link{N}} individuals.
}
\details{
Definition: Individuals for which
\code{cond.true} and \code{dec.neg}
holds simultaneously.

\code{mi} are cases of incorrect
classifications (type-II errors).

Relationships:
In a population of size \code{\link{N}}
the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}}

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}}

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}}
}
}
\seealso{
\code{\link{mirt}} is the probability or rate of misses;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{is_freq}} verifies frequencies.

Other essential parameters: \code{\link{cr}},
  \code{\link{fa}}, \code{\link{hi}}

Other frequencies: \code{\link{N}}, \code{\link{cr}},
  \code{\link{fa}}, \code{\link{hi}}
}
\keyword{datasets}
