% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mlm.R
\name{mlm_trans}
\alias{mlm_trans}
\title{Transform a 2x2 matrix (into a table of probabilities/conditional probabilities).}
\usage{
mlm_trans(mx, margin = 0, as_pc = FALSE, n_digits = 3)
}
\arguments{
\item{mx}{A 2x2 matrix (as numeric contingency table, required).}

\item{margin}{Margin to conditionalize table values (numeric, from 0 to 3).
Default: \code{margin = 0} (i.e., unconditional probabilities).}

\item{as_pc}{Boolean: Convert probabilities into percentages?
Default: \code{as_pc = FALSE}.}

\item{n_digits}{Number of decimal places to which result is rounded.
Default: \code{n_digits = 3}.}
}
\value{
A numeric 2x2 table.
}
\description{
\code{mlm_trans} converts a 2x2 matrix \code{mx}
(i.e., a numeric contingency table of frequency counts)
into a table of probabilities or marginal probabilities.
}
\details{
By default, \code{mlm_trans} converts a 2x2 matrix of frequency counts
into the corresponding probability values (\code{margin = 0}).

Setting \code{margin} to 1 to 3 computes
conditional probabilities in a
by row (\code{margin = 1}),
by column (\code{margin = 2}), or
by diagonal (\code{margin = 3}) fashion.

The tables for \code{margin = 0} to \code{margin = 2}
are computed by the \strong{base} R function \code{\link{prop.table}}.
}
\examples{
# The mammography problem:
abcd <- c(8, 95, 2, 895)  # Frequencies (Gigerenzer & Hoffrage, 1995)
mp <- mlm_frame(data = abcd, x = "Condition", y = "Test",
                x_levels = c("cancer", "no cancer"),
                y_levels = c("positive", "negative"))

mlm_trans(mp)
mlm_trans(mp, margin = 0)  # by-cell: probabilities
mlm_trans(mp, margin = 1)  # by-row: conditional prob
mlm_trans(mp, margin = 2)  # by-col
mlm_trans(mp, margin = 3)  # by-diagonals

mlm_trans(mp, as_pc = TRUE, n_digits = 2)  # as percentages

# The following must sum to 1:
sum(mlm_trans(mp, margin = 0))      # 4 cell values
rowSums(mlm_trans(mp, margin = 1))  # 2 row sums
colSums(mlm_trans(mp, margin = 2))  # 2 col sums


}
\seealso{
\code{\link{mlm_frame}} creates a 2x2 matrix;
\code{\link{mlm_focus}} computes measures for a 2x2 matrix;
\code{\link{is_matrix}} verifies a 2x2 matrix.

Other matrix lens model functions: 
\code{\link{mlm_focus}()},
\code{\link{mlm_frame}()}
}
\concept{matrix lens model functions}
