% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prob.R
\docType{data}
\name{NPV}
\alias{NPV}
\title{The negative predictive value of a decision process or diagnostic procedure.}
\format{An object of class \code{logical} of length 1.}
\usage{
NPV
}
\description{
\code{NPV} defines some decision's negative predictive value (NPV):
The conditional probability of the condition being FALSE
provided that the decision is negative.
}
\details{
Understanding or obtaining the negative predictive value \code{NPV}:

\itemize{

\item Definition:
  \code{NPV} is the conditional probability
  for the condition being \code{FALSE}
  given a negative decision:

\code{NPV = p(condition = FALSE | decision = negative)}

or the probability of a negative decision being correct.

\item Relationships:

a. \code{NPV} is the complement of the
  false omission rate \code{\link{FOR}}:

\code{NPV = 1 - FOR}

b. \code{NPV} is the opposite conditional probability
  -- but not the complement --
  of the specificity \code{\link{spec}}:

\code{spec = p(decision = negative | condition = FALSE)}

\item In natural frequencies,
  \code{NPV} is the ratio of individuals for which
  \code{decision = negative} and \code{condition = FALSE}
  divided by the number of all individuals for which
  \code{decision = negative}:

\code{NPV = n(decision = negative & condition = FALSE) / n(decision = negative)}

\code{NPV = n.cr / (n.mi + n.cr)}

\item \code{NPV} is a feature of a decision process
  or diagnostic procedure and
  -- similar to the specificity \code{\link{spec}} --
  a measure of correct decisions (negative decisions
  that are actually FALSE).

However, due to being a conditional probability,
  the value of \code{NPV} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
NPV <- .95     # => sets a negative predictive value of 95\%
NPV <- 95/100  # => (condition = FALSE) for 95 people out of 100 people for which (decision = negative)
is_prob(NPV)   # => TRUE (as NPV is a probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values}{Wikipedia} for additional information.
}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{comp_NPV}} and \code{\link{comp_NPV_freq}} compute PPVs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities.

Other functions computing probabilities: \code{\link{FDR}},
  \code{\link{FOR}}, \code{\link{PPV}},
  \code{\link{comp_comp_pair}}, \code{\link{comp_fart}},
  \code{\link{comp_mirt}}, \code{\link{comp_prob_comp}},
  \code{\link{comp_prob}}, \code{\link{comp_sens}},
  \code{\link{comp_spec}}
}
\keyword{datasets}
