% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pal.R
\docType{data}
\name{pal_gbs}
\alias{pal_gbs}
\title{Alternative color palette with gbs (green/blue/sand) colors.}
\format{An object of class \code{character} of length 15.}
\usage{
pal_gbs
}
\description{
\code{pal_gbs} is initialized to a vector of named elements (colors)
to define an alternative gbs (green/blue/sand) scenario color scheme.
}
\details{
See \code{\link{pal}} for default color information.

Assign \code{pal <- pal_gbs} to use as default color scheme
throughout the \code{riskyr} package.
}
\examples{
pal_gbs        # shows all current color names and values
pal_gbs["hi"]  # shows the current color for hits (true positives)
pal_gbs["hi"] <- "gold" # defines a new color for hits (true positives, TP)

}
\seealso{
\code{\link{pal}} contains current color information;
\code{\link{init_pal}} initializes color information.

Other lists containing current scenario information: \code{\link{accu}},
  \code{\link{freq}}, \code{\link{num}},
  \code{\link{pal_4c}}, \code{\link{pal_bw}},
  \code{\link{pal_kn}}, \code{\link{pal_org}},
  \code{\link{pal_vir}}, \code{\link{pal}},
  \code{\link{prob}}, \code{\link{txt_TF}},
  \code{\link{txt}}
}
\concept{lists containing current scenario information}
\keyword{datasets}
