% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pal.R
\docType{data}
\name{pal_unikn}
\alias{pal_unikn}
\title{Alternative color palette for unikn.}
\format{
An object of class \code{character} of length 16.
}
\usage{
pal_unikn
}
\description{
\code{pal_unikn} is initialized to a vector of named elements (colors)
to define an alternative (unikn) scenario color scheme.
}
\details{
See \code{\link{pal}} for default color information.

Assign \code{pal <- pal_unikn} to use as default color scheme
throughout the \bold{riskyr} package.
}
\examples{
pal_unikn        # shows all color names and current values
pal_unikn["hi"]  # shows the current color for hits (true positives, TP)
pal_unikn["hi"] <- "grey" # defines a new color for hits (true positives, TP)

}
\seealso{
\code{\link{pal_kn}} contains fewer \bold{unikn} colors;
\code{\link{pal}} contains current color information;
\code{\link{init_pal}} initializes color information.

Other lists containing current scenario information: 
\code{\link{accu}},
\code{\link{freq}},
\code{\link{num}},
\code{\link{pal_bwp}},
\code{\link{pal_bw}},
\code{\link{pal_kn}},
\code{\link{pal_mbw}},
\code{\link{pal_mod}},
\code{\link{pal_org}},
\code{\link{pal_rgb}},
\code{\link{pal_vir}},
\code{\link{pal}},
\code{\link{prob}},
\code{\link{txt_TF}},
\code{\link{txt_org}},
\code{\link{txt}}
}
\concept{lists containing current scenario information}
\keyword{datasets}
