% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{plot.box}
\alias{plot.box}
\title{Plot information of a riskyr object.}
\usage{
\method{plot}{box}(obj, cur_freq = freq, cur_txt = txt,
  cur_pal = pal, ...)
}
\arguments{
\item{obj}{The object of class "box" to be plotted.}

\item{cur_freq}{The frequency information related to this box object.}

\item{cur_txt}{The text information associate with this box object.}

\item{cur_pal}{The color information related to this box object. #'}

\item{...}{Additional parameters to be passed to the
underlying plotting functions.}
}
\description{
\code{plot.box} is a utility method that allows to plot low level boxes for \code{riskyr} plots.
}
\details{
\code{plot.riskyr} also uses the text settings
specified in the "riskyr" object.
}
\seealso{
Other visualization functions: \code{\link{plot.riskyr}},
  \code{\link{plot_area}}, \code{\link{plot_bar}},
  \code{\link{plot_curve}}, \code{\link{plot_fnet}},
  \code{\link{plot_icons}}, \code{\link{plot_mosaic}},
  \code{\link{plot_plane}}, \code{\link{plot_prism}},
  \code{\link{plot_tab}}, \code{\link{plot_tree}}
}
\concept{visualization functions}
