% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_risk.R
\name{plot_cbar}
\alias{plot_cbar}
\title{Plot cumulative risk dynamics (as bar chart)}
\usage{
plot_cbar(
  r = 0.5,
  t = NA,
  N = 100,
  horizontal = TRUE,
  sort = FALSE,
  N_max = 100,
  bar_width = 0.5,
  show_trans = 1,
  show_ev = TRUE,
  show_n = FALSE,
  show_bin = FALSE,
  colors = c("firebrick", "grey96", "green4", "grey40")
)
}
\arguments{
\item{r}{risk (probability of occurrence per time period).
A non-scalar vector allows for different risk values
at different times (and \code{t = length(r)}).}

\item{t}{time periods/rounds.
Default: \code{t = NA}, setting \code{t <- length(r)}.
A scalar \code{r} and numeric \code{t} sets \code{r <- rep(r, t)}.}

\item{N}{population size.
Default: \code{N = 100} expresses risks as percentages,
\code{N = 1} as probabilities, else frequencies.}

\item{horizontal}{logical: Draw horizontal vs. vertical bars?}

\item{sort}{logical: Sort outputs by number of event occurrences?
Default: \code{sort = FALSE}.}

\item{N_max}{maximum N value plotted (for zooming in for small \code{r} values).
Default value should be set to population size \code{N}.}

\item{bar_width}{width of (horizontal/vertical) bar per time period.
Default: \code{bar_width = .50}.}

\item{show_trans}{numeric: Show transition polygons (between bars)?
Values of 0/1/2/3 focus on no/new/remaining/both risk segments, respectively.
Default: \code{show_trans = 1} (i.e., focus on increasing risk segments).}

\item{show_ev}{logical: Show number of risky event occurrence (as bar label)?
Default: \code{show_ev = TRUE}.}

\item{show_n}{logical: Show population frequency of risky event occurrences (as bar label)?
Default: \code{show_n = FALSE}.}

\item{show_bin}{logical: Show risky event history as binary state representation (as bar label)?
Default: \code{show_bin = FALSE}.}

\item{colors}{A vector of color values
(for risk event frequency being "hi", "lo", "no", and "bd" borders, respectively).
Default: \code{colors = c("firebrick", "grey96", "green4", "grey40")}.}
}
\value{
data of p-values, named by number of event occurrences
(invisibly, as list of named vectors, for each time period t).
}
\description{
\code{plot_cbar} plots the results of cumulative risk dynamics
as a bar chart (with percentages of risk event counts
for each period t on a horizontal bar).
}
