% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crisk.R
\name{plot_crisk}
\alias{plot_crisk}
\title{Plot a cumulative risk curve.}
\usage{
plot_crisk(
  x,
  y,
  x_from = NA,
  x_to = NA,
  fit_curve = FALSE,
  show_inc = TRUE,
  show_pas = TRUE,
  show_rem = TRUE,
  show_aux = TRUE,
  prev = num$prev,
  sens = num$sens,
  mirt = NA,
  spec = num$spec,
  fart = NA,
  N = num$N,
  by = "cddc",
  p_split = "v",
  area = "sq",
  scale = "p",
  round = TRUE,
  sample = FALSE,
  sum_w = 0.1,
  gaps = c(NA, NA),
  f_lbl = "num",
  f_lbl_sep = NA,
  f_lbl_sum = "num",
  f_lbl_hd = "nam",
  f_lwd = 0,
  p_lbl = NA,
  arr_c = -3,
  col_p = c(grey(0.15, 0.99), "yellow", "yellow"),
  brd_dis = 0.06,
  lbl_txt = txt,
  title_lbl = "",
  cex_lbl = 0.9,
  cex_p_lbl = NA,
  col_pal = pal_crisk,
  mar_notes = TRUE,
  ...
)
}
\arguments{
\item{x}{Values on an x-dimension on which risk is expressed
(required, as a vector).}

\item{y}{Values of cumulative risks on an y-dimension
(required, as a vector),
as monotonically increasing percentage values
(0 <= y <= 100).
Pairs of \code{x} and \code{y} are assumed to
correspond to each other.}

\item{x_from}{Start value of risk increment.}

\item{x_to}{End value of risk increment.}

\item{fit_curve}{Boolean: Fit a curve to \code{x}-\code{y}-data?
Default: \code{fit_curve = FALSE}.}

\item{show_inc}{Boolean: Show risk increments?
Default: \code{show_inc = TRUE}.}

\item{show_pas}{Boolean: Show past/passed risk?
Default: \code{show_pas = TRUE}.}

\item{show_rem}{Boolean: Show remaining risk?
Default: \code{show_rem = TRUE}.}

\item{show_aux}{Boolean: Show auxiliary elements (lines and points)?
Default: \code{show_aux = TRUE}.}

\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the population.
A suitable value of \code{\link{N}} is computed, if not provided.
Note: \code{\link{N}} is not represented in the plot,
but used for computing frequency information \code{\link{freq}}
from current probabilities \code{\link{prob}}.}

\item{by}{A character code specifying 2 perspectives that split the population into subsets,
with 6 options:
  \enumerate{
  \item \code{"cddc"}: by condition (cd) and by decision (dc) (default);
  \item \code{"cdac"}: by condition (cd) and by accuracy (ac);
  \item \code{"dccd"}: by decision (dc) and by condition (cd);
  \item \code{"dcac"}: by decision (dc) and by accuracy (ac);
  \item \code{"accd"}: by accuracy (ac) and by condition (cd);
  \item \code{"acdc"}: by accuracy (ac) and by decision (dc).
  }}

\item{p_split}{Primary perspective for population split,
with 2 options:
  \enumerate{
  \item \code{"v"}: vertical (default);
  \item \code{"h"}: horizontal.
  }}

\item{area}{A character code specifying the shape of the main area,
with 2 options:
  \enumerate{
  \item \code{"sq"}: main area is scaled to square (default);
  \item \code{"no"}: no scaling (rectangular area fills plot size).
  }}

\item{scale}{Scale probabilities and corresponding area dimensions either by
exact probability or by (rounded or non-rounded) frequency, with 2 options:
  \enumerate{
  \item \code{"p"}: scale main area dimensions by exact probability (default);
  \item \code{"f"}: re-compute probabilities from (rounded or non-rounded) frequencies
  and scale main area dimensions by their frequency.
  }
 Note: \code{scale} setting matters for the display of probability values and for
 area plots with small population sizes \code{\link{N}} when \code{round = TRUE}.}

\item{round}{A Boolean option specifying whether computed frequencies
are rounded to integers. Default: \code{round = TRUE}.}

\item{sample}{Boolean value that determines whether frequency values
are sampled from \code{N}, given the probability values of
\code{prev}, \code{sens}, and \code{spec}.
Default: \code{sample = FALSE}.}

\item{sum_w}{Border width of 2 perspective summaries
(on top and left borders) of main area as a proportion of area size
(i.e., in range \code{0 <= sum_w <= 1}).
Default: \code{sum_w = .10}.
Setting \code{sum_w = 0}, \code{NA}, or \code{NULL} removes summaries;
setting \code{sum_w = 1} scales summaries to same size as main areas.}

\item{gaps}{Size of gaps (as binary numeric vector) specifying
the width of vertical and horizontal gaps as proportions of area size.
Defaults: \code{gaps = c(.02, .00)} for \code{p_split = "v"} and
\code{gaps = c(.00, .02)} for \code{p_split = "h"}.}

\item{f_lbl}{Type of label for showing frequency values in 4 main areas,
with 6 options:
  \enumerate{
  \item \code{"def"}: abbreviated names and frequency values;
  \item \code{"abb"}: abbreviated frequency names only (as specified in code);
  \item \code{"nam"}: names only (as specified in \code{lbl_txt = txt});
  \item \code{"num"}: numeric frequency values only (default);
  \item \code{"namnum"}: names (as specified in \code{lbl_txt = txt}) and numeric values;
  \item \code{"no"}: no frequency labels (same for \code{f_lbl = NA} or \code{NULL}).
  }}

\item{f_lbl_sep}{Label separator for main frequencies
(used for \code{f_lbl = "def" OR "namnum"}).
Use \code{f_lbl_sep = ":\n"} to add a line break between name and numeric value.
Default: \code{f_lbl_sep = NA} (set to \code{" = "} or \code{":\n"} based on \code{f_lbl}).}

\item{f_lbl_sum}{Type of label for showing frequency values in summary cells,
with same 6 options as \code{f_lbl} (above).
Default: \code{f_lbl_sum = "num"}: numeric values only.}

\item{f_lbl_hd}{Type of label for showing frequency values in header,
with same 6 options as \code{f_lbl} (above).
Default: \code{f_lbl_hd = "nam"}: names only (as specified in \code{lbl_txt = txt}).}

\item{f_lwd}{Line width of areas.
Default: \code{f_lwd = 0}.}

\item{p_lbl}{Type of label for showing 3 key probability links and values,
with 7 options:
  \enumerate{
  \item \code{"def"}: show links and abbreviated names and probability values;
  \item \code{"abb"}: show links and abbreviated probability names;
  \item \code{"nam"}: show links and probability names (as specified in code);
  \item \code{"num"}: show links and numeric probability values;
  \item \code{"namnum"}: show links with names and numeric probability values;
  \item \code{"no"}: show links with no labels;
  \item \code{NA}: show no labels or links (same for \code{p_lbl = NULL}, default).
  }}

\item{arr_c}{Arrow code for symbols at ends of probability links
(as a numeric value \code{-3 <= arr_c <= +6}),
with the following options:
  \itemize{
  \item \code{-1} to \code{-3}: points at one/other/both end/s;
  \item \code{0}: no symbols;
  \item \code{+1} to \code{+3}: V-arrow at one/other/both end/s;
  \item \code{+4} to \code{+6}: T-arrow at one/other/both end/s.
}
Default: \code{arr_c = -3} (points at both ends).}

\item{col_p}{Colors of probability links (as vector of 3 colors).
Default: \code{col_p = c(grey(.15, .99), "yellow", "yellow")}.
(Also consider: "black", "cornsilk", "whitesmoke").}

\item{brd_dis}{Distance of probability links from area border
(as proportion of area width).
Default: \code{brd_dis = .06}.
Note: Adjust to avoid overlapping labels.
Negative values show links outside of main area.}

\item{lbl_txt}{Default label set for text elements.
Default: \code{lbl_txt = \link{txt}}.}

\item{title_lbl}{Text label for current plot title.
Default: \code{title_lbl = ""} (using "Cumulative risk").}

\item{cex_lbl}{Scaling factor for text labels (frequencies and headers).
Default: \code{cex_lbl = .90}.}

\item{cex_p_lbl}{Scaling factor for text labels (probabilities).
Default: \code{cex_p_lbl = cex_lbl - .05}.}

\item{col_pal}{Color palette.
Default: \code{col_pal = \link{pal}}.}

\item{mar_notes}{Boolean option for showing margin notes.
Default: \code{mar_notes = TRUE}.}

\item{...}{Other (graphical) parameters.}
}
\value{
Nothing (NULL).
}
\description{
\code{plot_crisk} creates visualizations of
cumulative risks.
}
\details{
Assumes data inputs \code{x} and \code{y},
which correspond to each other so that
\code{y} is a (monotonically increasing)
probability density function (over risk increments
expressed as a function of \code{x}).
}
\examples{
# plot_crisk(x = seq(0, 100, by = 10),
#            y = c(0, 0, 0, 10, 30, 50, 70, 80, 0, 0, 0))

}
\seealso{
Other visualization functions: 
\code{\link{plot.riskyr}()},
\code{\link{plot_area}()},
\code{\link{plot_bar}()},
\code{\link{plot_curve}()},
\code{\link{plot_fnet}()},
\code{\link{plot_icons}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_plane}()},
\code{\link{plot_prism}()},
\code{\link{plot_tab}()},
\code{\link{plot_tree}()}
}
\concept{visualization functions}
