% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_risk.R
\name{plot_cum_bar}
\alias{plot_cum_bar}
\title{plot_cum_bar plots the results of cumulative risk computations
as a bar plot (with percentages of risk event counts
for each period t as a horizontal bar).}
\usage{
plot_cum_bar(data, N = 100)
}
\arguments{
\item{data}{The data to plot (as list of named probabilities).}

\item{N}{population size.}
}
\description{
plot_cum_bar plots the results of cumulative risk computations
as a bar plot (with percentages of risk event counts
for each period t as a horizontal bar).
}
