% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iconarray.R
\name{plot_iconarray}
\alias{plot_iconarray}
\title{Plot an icon array of a population.}
\usage{
plot_iconarray(prev = num$prev, sens = num$sens, mirt = NA,
  spec = num$spec, fart = NA, N = freq$N, type.sort = "mosaic",
  ident.order = c("hi", "mi", "fa", "cr"), random.position = FALSE,
  random.identities = FALSE, icon.colors = pal[c("hi", "mi", "fa", "cr")],
  icon.types = 22, pch.border = grey(0.1, 0.5), pch.lwd = 1.5,
  transparency = 0.5, block.d = NULL, border.d = 0.1,
  block_size_col = 10, block_size_row = 10, ncol_blocks = NULL,
  nrow_blocks = NULL, fill_array = "left", fill_blocks = "rowwise",
  title.lbl = txt$scen.lbl, xlim = c(0, 1), ylim = c(0, 1), cex = NULL,
  ...)
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the population.
A suitable value of \code{\link{N}} is computed, if not provided.}

\item{type.sort}{Bla bla bla... with the following options:

\enumerate{

  \item \code{type.sort = "mosaic"} Order in some way.

  \item \code{type.sort = "xyz"} Order in some other way.#'

}}

\item{ident.order}{Specifies the order in which icon identities
(hits, misses, false alarms, and correct rejections) are plotted.
Default: \code{ident.order = c("hi", "mi", "fa", "cr")}}

\item{random.position}{Are positions randomly drawn?
Default: \code{random.position = FALSE}.}

\item{random.identities}{Are identities randomly assigned to positions?
Default: \code{random.identities = FALSE}


Various other options allow the customization of text labels and colors:}

\item{title.lbl}{Text label to set plot title.}
}
\description{
\code{plot_iconarray} draws an icon array of
a population of which individual's condition has been classified
correctly or incorrectly
from a sufficient and valid set of 3 essential probabilities
(\code{\link{prev}}, and
\code{\link{sens}} or its complement \code{\link{mirt}}, and
\code{\link{spec}} or its complement \code{\link{fart}})
or existing frequency information \code{\link{freq}}
and a population size of \code{\link{N}} individuals.
}
\details{
If probabilities are provided, a new list of
natural frequencies \code{\link{freq}} is computed by \code{\link{comp_freq}}.
By contrast, if no probabilities are provided,
the values currently contained in \code{\link{freq}} are used.
By default, \code{\link{comp_freq}} rounds frequencies to nearest integers
to avoid decimal values in \code{\link{freq}}.
}
\examples{
# ways to work:
plot_iconarray()  # => plots icon array for default population

# standard icon arrays:
plot_iconarray(N = 800, icon.types = c(21,23,24,23), block.d = 0.5, border.d = 0.5)

plot_iconarray(N = 800, icon.types = c(21,23,24,23), block.d = 0.5, border.d = 0.5,
               random.identities = TRUE)

plot_iconarray(N = 10000, icon.types = c(21,23,24,23), block.d = 0.5, border.d = 0.5)

plot_iconarray(N = 10000, icon.types = c(21,23,24,23),
               random.identities = TRUE)

plot_iconarray(icon.types = c(21,23,24,23),
               block_size_col = 5, block_size_row = 5, #ncol_blocks = 2, nrow_blocks = 2,
               block.d = 0.5, border.d = 0.9)

plot_iconarray(N = 1000, icon.types = c(21,23,24,23), block.d = 0.4)

plot_iconarray(N = 1250, sens = 0.9, spec = 0.9, prev = 0.9,
               icon.types = c(21,23,24,23),
               block_size_col = 10, block_size_row = 5,
               ncol_blocks = 5, nrow_blocks = 5,
               block.d = 0.8,
               border.d = 0.2,
               fill_array = "top")


plot_iconarray(N = 10000, sens = 0.9, spec = 0.6, prev = 0.3,
               icon.types = c(21,23,21,23),
               ident.order = c("hi", "mi", "cr", "fa"),
               block.d = 0.8,
               border.d = 0.01,
               cex = 0.7,
               random.position = FALSE,
               random.identities = FALSE)


# Mosaic like and randomized arrays:
plot_iconarray(N = 1000, icon.types = c(22,23,22,23), #cex = 3,
               random.position = TRUE, type.sort = "mosaic", block.d = 0.05)

plot_iconarray(icon.types = c(22,23,21,23), #cex = 10,
               ident.order = c("mi", "hi", "cr", "fa"),
               random.position = TRUE, type.sort = "equal", block.d = 0.05)

plot_iconarray(icon.types = c(21,23,22,23), #cex = 10,
               random.position = TRUE, random.identities = TRUE)

plot_iconarray(N = 10000, sens = 0.9, spec = 0.9, prev = 0.9,
               icon.types = c(21,23,21,23),
               ident.order = c("hi", "mi", "cr", "fa"),
               block.d = 0.1,
               border.d = 0.01,
               cex = 0.7,
               random.position = TRUE,
               random.identities = FALSE,
               type.sort = "mosaic")


}
\seealso{
Other visualization functions: \code{\link{plot_PV3d}},
  \code{\link{plot_PV}}, \code{\link{plot_fnet}},
  \code{\link{plot_mosaic}}, \code{\link{plot_tree}}
}
