% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mosaic.R
\name{plot_mosaic}
\alias{plot_mosaic}
\title{Plot a mosaic plot of population frequencies.}
\usage{
plot_mosaic(pop = popu, vsplit = TRUE, title.lbl = txt$scen.lbl,
  col.sdt = pal[4:7])
}
\description{
\code{plot_mosaic} visualizes a mosaic plot that
represents the proportions of frequencies in the current
population \code{\link{popu}} as relatives sizes of
rectangular areas.
}
\details{
Corresponding rectangles can be visualized by
opting for vertical rectangles (by selecting
the option \code{box = "vr"}) in \code{\link{plot_tree}}.

\code{plot_mosaic} requires and uses the R packages "vcd" and
"grid" (\code{library("vcd", "grid")}).
}
\examples{
plot_mosaic()
plot_mosaic(title.lbl = "")
plot_mosaic(vsplit = FALSE)
plot_mosaic(title.lbl = "My favorite scenario", col.sdt = "goldenrod")

}
\seealso{
The corresponding data frame \code{\link{popu}}
is generated by \code{\link{comp_popu}};
\code{\link{num}} for basic numeric parameters;
\code{\link{txt}} for current text settings;
\code{\link{pal}} for current color settings
}
