% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PV.R
\name{plot_PV}
\alias{plot_PV}
\title{Plot predictive values (PPV and NPV) as a function of prevalence.}
\usage{
plot_PV(prev = num$prev, sens = num$sens, mirt = NA, spec = num$spec,
  fart = NA, show.PVprev = TRUE, show.PVpoints = TRUE,
  log.scale = FALSE, title.lbl = txt$scen.lbl, col.ppv = pal["ppv"],
  col.npv = pal["npv"])
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{show.PVprev}{Option for showing current
prevalence \code{\link{prev}} in the plot (as a vertical line).
Default: \code{show.PVprev = TRUE}.}

\item{show.PVpoints}{Option for showing current
predictive values (\code{\link{PPV}} and \code{\link{NPV}})
on the corresponding curve.
Default: \code{show.PVpoints = TRUE}.}

\item{log.scale}{Boolean value for switching from a linear
to a logarithmic x-axis. Default: \code{log.scale = FALSE}.}

\item{title.lbl}{The title of the current plot.}

\item{col.ppv}{The color in which PPV is shown.}
}
\description{
\code{plot_PV} draws curves that illustrate
the positive predictive value (\code{\link{PPV}})
and negative predictive value (\code{\link{NPV}})
as a function of the prevalence (\code{\link{prev}}).
}
\examples{
plot_PV()
plot_PV(prev = .0001, log.scale = TRUE)
plot_PV(title.lbl = "My favorite scenario", show.PVprev = FALSE)
plot_PV(title.lbl = "", col.ppv = "firebrick3", col.npv = "steelblue3")


}
\seealso{
\code{\link{comp_popu}} computes the current population;
\code{\link{popu}} contains the current population;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{freq}} contains current frequency information;
\code{\link{num}} for basic numeric parameters;
\code{\link{txt}} for current text settings;
\code{\link{pal}} for current color settings

Other visualization functions: \code{\link{plot_PV3d}},
  \code{\link{plot_fnet}}, \code{\link{plot_iconarray}},
  \code{\link{plot_mosaic}}, \code{\link{plot_tree}}
}
