% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PV3d.R
\name{plot_PV3d}
\alias{plot_PV3d}
\title{Plot predictive values (PPV or NPV) as a function of
sensitivity and specificity.}
\usage{
plot_PV3d(prev = num$prev, sens = num$sens, mirt = NA, spec = num$spec,
  fart = NA, is.ppv = TRUE, step.size = 0.05, show.PVpoints = TRUE,
  cur.theta = -45, cur.phi = 0, cur.d = 1.5, cur.expand = 1.1,
  cur.ltheta = 200, cur.shade = 0.25, title.lbl = txt$scen.lbl,
  col.pv = pal["ppv"])
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{is.ppv}{Option for showing either
\code{\link{PPV}} or \code{\link{NPV}}.
Default: \code{is.ppv = TRUE}.}

\item{step.size}{Option for the step size of the range of
\code{\link{sens}} and \code{\link{spec}} values shown.
Default: \code{step.size = .05}.}

\item{show.PVpoints}{Option for showing current
\code{\link{PPV}} or \code{\link{NPV}})
on the corresponding plane.
Default: \code{show.PVpoints = TRUE}.}

\item{cur.theta}{Horizontal rotation angle (used by \code{\link{persp}}).
Default: \code{cur.theta = -45}.}

\item{cur.phi}{Vertical rotation angle (used by \code{\link{persp}}).
Default: \code{cur.phi = 0}.}

\item{title.lbl}{The title of the current plot.}

\item{col.pv}{The color in which the selected predictive value is shown.}
}
\description{
\code{plot_PV3d} draws a plane that illustrates
either the positive predictive value (\code{\link{PPV}})
or the negative predictive value (\code{\link{NPV}})
as a function of
a decision's sensitivity \code{\link{sens}} and
a decision's specificity value \code{\link{spec}}
for a given prevalence (\code{\link{prev}}).
}
\examples{
plot_PV3d()  # => shows PPV plane (using current defaults)
plot_PV3d(prev = .5, show.PVpoints = FALSE, step.size = .5, title.lbl = "Quick test")
plot_PV3d(prev = .3, is.ppv = FALSE, col.pv = pal["npv"])
plot_PV3d(prev = .5, is.ppv = FALSE, step.size = .20, title.lbl = "",
          cur.theta = -45, cur.phi = 45, cur.expand = 1.4, col.pv = "firebrick3")


}
\seealso{
\code{\link{comp_popu}} computes the current population;
\code{\link{popu}} contains the current population;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{freq}} contains current frequency information;
\code{\link{num}} for basic numeric parameters;
\code{\link{txt}} for current text settings;
\code{\link{pal}} for current color settings

Other visualization functions: \code{\link{plot_PV}},
  \code{\link{plot_fnet}}, \code{\link{plot_mosaic}},
  \code{\link{plot_tree}}
}
