% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tab.R
\name{plot_tab}
\alias{plot_tab}
\title{Plot a 2 x 2 contingency table of population frequencies.}
\usage{
plot_tab(prev = num$prev, sens = num$sens, mirt = NA,
  spec = num$spec, fart = NA, N = num$N, by = "cddc",
  p_split = "v", area = "no", scale = "p", round = TRUE,
  f_lbl = "num", f_lbl_sep = " = ", f_lbl_sum = f_lbl,
  f_lbl_hd = "abb", f_lwd = 0, gaps = c(NA, NA), brd_w = 0.1,
  p_lbl = NA, arr_c = -3, col_p = c(grey(0.15, 0.99), "yellow",
  "yellow"), brd_dis = 0.3, lbl_txt = txt, title_lbl = txt$scen_lbl,
  cex_lbl = 0.9, cex_p_lbl = NA, col_pal = pal, mar_notes = TRUE,
  ...)
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the population.
A suitable value of \code{\link{N}} is computed, if not provided.
Note: \code{\link{N}} is not represented in the plot,
but used for computing frequency information \code{\link{freq}}
from current probabilities \code{\link{prob}}.}

\item{by}{A character code specifying 2 perspectives that split the population into subsets,
with 6 options:
  \enumerate{
  \item \code{"cddc"}: by condition (cd) and by decision (dc) (default);
  \item \code{"cdac"}: by condition (cd) and by accuracy (ac);
  \item \code{"dccd"}: by decision (dc) and by condition (cd);
  \item \code{"dcac"}: by decision (dc) and by accuracy (ac);
  \item \code{"accd"}: by accuracy (ac) and by condition (cd);
  \item \code{"acdc"}: by accuracy (ac) and by decision (dc).
  }}

\item{p_split}{Primary perspective for population split,
with 2 options:
  \enumerate{
  \item \code{"v"}: vertical (default);
  \item \code{"h"}: horizontal.
  }
Note: In contrast to \code{\link{plot_area}}, this setting only
determines which 3 probability links are shown
(e.g., when \code{p_link = "def"}).}

\item{area}{A character code specifying the shape of the main area,
with 4 options:
  \enumerate{
  \item \code{"sq"}: main area is scaled to square;
  \item \code{"no"}: no scaling (rectangular area fills plot size; default).
  }}

\item{scale}{Scale probabilities (but not table cell dimensions) either by
exact probability or by (rounded or non-rounded) frequency, with 2 options:
  \enumerate{
  \item \code{"p"}: scale main area dimensions by exact probability (default);
  \item \code{"f"}: re-compute probabilities from (rounded or non-rounded) frequencies
  and scale main area dimensions by their frequency.
  }
 Note: \code{scale} setting matters for the display of probability values and for
 area plots with small population sizes \code{\link{N}} when \code{round = TRUE}.}

\item{round}{A Boolean option specifying whether computed frequencies
are rounded to integers. Default: \code{round = TRUE}.}

\item{f_lbl}{Type of label for showing frequency values in 4 main areas,
with 6 options:
  \enumerate{
  \item \code{"def"}: abbreviated names and frequency values (default);
  \item \code{"abb"}: abbreviated frequency names only (as specified in code);
  \item \code{"nam"}: names only (as specified in \code{lbl_txt = txt});
  \item \code{"num"}: numeric frequency values only;
  \item \code{"namnum"}: names (as specified in \code{lbl_txt = txt}) and numeric values;
  \item \code{"no"}: no frequency labels (same for \code{f_lbl = NA} or \code{NULL}).
  }}

\item{f_lbl_sep}{Label separator for main frequencies
(when \code{f_lbl = "def" OR "namnum"}).
Use \code{f_lbl_sep = ":\n"} to add extra line between name and numeric value.
Default: \code{f_lbl_sep = " = "}.}

\item{f_lbl_sum}{Type of label for showing frequency values in summary cells,
with same 6 options as \code{f_lbl} (above).
Default: \code{f_lbl_sum = "def"}: abbreviated names and numeric values.}

\item{f_lbl_hd}{Type of label for showing frequency values in header,
with same 6 options as \code{f_lbl} (above).
Default: \code{f_lbl_hd = "abb"}: abbreviated names only.}

\item{f_lwd}{Line width of areas.
Default: \code{f_lwd = 1}.}

\item{gaps}{Size of gaps (as binary numeric vector) specifying
the widths of vertical and horizontal gaps between 2 x 2 table
and sums (in bottom row and right column).
Default: \code{gaps = c(.05, .06)}.}

\item{brd_w}{Border width for showing 2 perspective summaries
on top and left borders of main area (as a proportion of area size)
in a range \code{0 <= brd_w <= 1}.
Default: \code{brd_w = .10}.}

\item{p_lbl}{Type of label for showing 3 key probability links and values,
with 7 options:
  \enumerate{
  \item \code{"def"}: show links and abbreviated names and probability values;
  \item \code{"abb"}: show links and abbreviated probability names;
  \item \code{"nam"}: show links and probability names (as specified in code);
  \item \code{"num"}: show links and numeric probability values;
  \item \code{"namnum"}: show links with names and numeric probability values;
  \item \code{"no"}: show links with no labels;
  \item \code{NA}: no link (same for \code{p_lbl = NULL}, default).
  }}

\item{arr_c}{Arrow code for symbols at ends of probability links
(as a numeric value \code{-3 <= arr_c <= +6}),
with the following options:
  \itemize{
  \item \code{-1} to \code{-3}: points at one/other/both end/s;
  \item \code{0}: no symbols;
  \item \code{+1} to \code{+3}: V-arrow at one/other/both end/s;
  \item \code{+4} to \code{+6}: T-arrow at one/other/both end/s.
}
Default: \code{arr_c = -3} (points at both ends).}

\item{col_p}{Colors of probability links (as vector of 3 colors).
Default: \code{col_p = c(grey(.15, .99), "yellow", "yellow")}.}

\item{brd_dis}{Distance of probability links from cell center
(as a constant).
Default: \code{brd_dis = .30}.
Note: Adjust to avoid overlapping labels.}

\item{lbl_txt}{Default label set for text elements.
Default: \code{lbl_txt = \link{txt}}.}

\item{title_lbl}{Text label for current plot title.
Default: \code{title_lbl = txt$scen_lbl}.}

\item{cex_lbl}{Scaling factor for text labels (frequencies and headers).
Default: \code{cex_lbl = .90}.}

\item{cex_p_lbl}{Scaling factor for text labels (probabilities).
Default: \code{cex_p_lbl = cex_lbl - .05}.}

\item{col_pal}{Color palette.
Default: \code{col_pal = \link{pal}}.}

\item{mar_notes}{Boolean option for showing margin notes.
Default: \code{mar_notes = TRUE}.}

\item{...}{Other (graphical) parameters.}
}
\value{
Nothing (NULL).
}
\description{
\code{plot_tab} plots a 2 x 2 contingency table
(aka. confusion table) of
4 classification cases (\code{\link{hi}}, \code{\link{mi}},
\code{\link{fa}}, \code{\link{cr}})
and corresponding row and column sums.
}
\details{
\code{plot_tab} computes its frequencies \code{\link{freq}}
from a sufficient and valid set of 3 essential probabilities
(\code{\link{prev}}, and
\code{\link{sens}} or its complement \code{\link{mirt}}, and
\code{\link{spec}} or its complement \code{\link{fart}})
or existing frequency information \code{\link{freq}}
and a population size of \code{\link{N}} individuals.

\code{plot_tab} is derived from \code{\link{plot_area}},
but does not scale the dimensions of table cells.
}
\examples{
## Basics:
plot_tab()
plot_tab(area = "sq", f_lwd = 3)
plot_tab(f_lbl = "namnum", f_lbl_sep = ":\\n", brd_w = .10, f_lwd = .5)

## Computing local freq and prob values:
plot_tab(prev = .5, sens = 4/5, spec = 3/5, N = 10, f_lwd = 1)

## Plot versions:
# by x p_split [yields (3 x 2) x 2] = 12 versions]:
plot_tab(by = "cddc", p_split = "v", p_lbl = "def")  # v01 (see v07)
plot_tab(by = "cdac", p_split = "v", p_lbl = "def")  # v02 (see v11)
plot_tab(by = "cddc", p_split = "h", p_lbl = "def")  # v03 (see v05)
plot_tab(by = "cdac", p_split = "h", p_lbl = "def")  # v04 (see v09)

plot_tab(by = "dccd", p_split = "v", p_lbl = "def")  # v05 (is v03 rotated)
plot_tab(by = "dcac", p_split = "v", p_lbl = "def")  # v06 (see v12)
plot_tab(by = "dccd", p_split = "h", p_lbl = "def")  # v07 (is v01 rotated)
plot_tab(by = "dcac", p_split = "h", p_lbl = "def")  # v08 (see v10)

plot_tab(by = "accd", p_split = "v", p_lbl = "def")  # v09 (is v04 rotated)
plot_tab(by = "acdc", p_split = "v", p_lbl = "def")  # v10 (is v08 rotated)
plot_tab(by = "accd", p_split = "h", p_lbl = "def")  # v11 (is v02 rotated)
plot_tab(by = "acdc", p_split = "h", p_lbl = "def")  # v12 (is v06 rotated)


## Explore labels and links:
plot_tab(f_lbl = "abb", p_lbl = NA)  # abbreviated labels, no probability links
plot_tab(f_lbl = "num", f_lbl_sum = "abb", p_lbl = "num", f_lbl_hd = "abb")
plot_tab(f_lbl = "abb", f_lbl_sum = "abb", p_lbl = "nam", f_lbl_hd = "nam")
plot_tab(f_lbl = "namnum", f_lbl_sep = ":\\n",
         f_lbl_sum = "namnum", f_lbl_hd = "nam", p_lbl = "namnum")

## Misc. options:
plot_tab(area = "sq")        # area: square
plot_tab(title_lbl = "")     # no titles
plot_tab(mar_notes = FALSE)  # no margin notes

plot_tab(by = "cddc", gaps = c(.08, .00), area = "sq")    # gaps
plot_tab(by = "cddc", gaps = c(.02, .08), p_split = "h")  # gaps

# Showing prob as lines:
plot_tab(prev = 1/4, sens = 6/7, spec = 3/5, N = 100,
         by = "cddc", p_split = "v", col_pal = pal_rgb,
         p_lbl = "def", brd_dis = .25, arr_c = -3)
plot_tab(prev = 1/3, sens = 6/7, spec = 3/4, N = 100, scale = "f",
         by = "cddc", p_split = "h", col_pal = pal_rgb,
         p_lbl = "namnum", brd_dis = .15, arr_c = +3)

## Custom text labels and colors:
plot_tab(prev = .5, sens = 4/5, spec = 3/5, N = 10,
         by = "cddc", p_split = "v", area = "sq",
         lbl_txt = txt_TF,  # custom text
         f_lbl = "namnum", f_lbl_sep = ":\\n", f_lbl_sum = "num", f_lbl_hd  = "nam",
         col_pal = pal_mod, f_lwd = 3)  # custom colors

plot_tab(prev = .5, sens = 3/5, spec = 4/5, N = 10,
         by = "cddc", p_split = "h", area = "sq",
         lbl_txt = txt_TF,  # custom text
         f_lbl = "namnum", f_lbl_sep = ":\\n", f_lbl_sum = "num", f_lbl_hd  = "nam",
         col_pal = pal_kn, f_lwd = 1)  # custom colors

## Note some differences to plot_area (i.e., area/mosaic plot):
#
# In plot_tab:
#
# (1) p_split does not matter (except for selecting different prob links):
plot_tab(by = "cddc", p_split = "v")  # v01 (see v07)
plot_tab(by = "cddc", p_split = "h")  # v03 (see v05)
#
# (2) scale does not matter for dimensions (which are constant),
#     BUT matters for values shown in prob links and on margins:
plot_tab(N = 5, prev = .3, sens = .9, spec = .5,
         by = "cddc", scale = "p", p_lbl = "def", round = TRUE)  # (a) exact prob values
plot_tab(N = 5, prev = .3, sens = .9, spec = .5,
         by = "cddc", scale = "f", p_lbl = "def", round = TRUE)  # (b) prob from rounded freq!
plot_tab(N = 5, prev = .3, sens = .9, spec = .5,
         by = "cddc", scale = "f", p_lbl = "def", round = FALSE) # (c) same values as (a)


}
\seealso{
\code{\link{plot_area}} for plotting mosaic plot (scaling area dimensions);
\code{\link{pal}} contains current color settings;
\code{\link{txt}} contains current text settings.

Other visualization functions: \code{\link{plot.riskyr}},
  \code{\link{plot_area}}, \code{\link{plot_bar}},
  \code{\link{plot_curve}}, \code{\link{plot_fnet}},
  \code{\link{plot_icons}}, \code{\link{plot_mosaic}},
  \code{\link{plot_plane}}, \code{\link{plot_prism}},
  \code{\link{plot_tree}}
}
\concept{visualization functions}
