% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a tree diagram of natural frequencies.}
\usage{
plot_tree(prev = num$prev, sens = num$sens, spec = num$spec, fart = NA,
  N = freq$N, area = "no", round = TRUE, title.lbl = txt$scen.lbl,
  popu.lbl = txt$popu.lbl, cond.lbl = txt$cond.lbl,
  cond.true.lbl = txt$cond.true.lbl, cond.false.lbl = txt$cond.false.lbl,
  dec.lbl = txt$dec.lbl, dec.true.lbl = txt$dec.true.lbl,
  dec.false.lbl = txt$dec.false.lbl, sdt.hi.lbl = txt$sdt.hi.lbl,
  sdt.mi.lbl = txt$sdt.mi.lbl, sdt.fa.lbl = txt$sdt.fa.lbl,
  sdt.cr.lbl = txt$sdt.cr.lbl, col.boxes = pal[1:7], col.txt = grey(0.01,
  alpha = 0.99), col.border = col.grey.4, col.shadow = col.sand.dark,
  cex.shadow = 0)
}
\arguments{
\item{prev}{The condition's prevalence value \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity value \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the population:
a suitable value of \code{\link{N}} is computed, if not provided.}

\item{area}{A character option for 4 different box sizes:
- \code{"no"} ... all boxes are shown with the same size (default);
- \code{"sq"} ... boxes as squares of area sizes proportional to frequencies;
- \code{"hr"} ... boxes as horizontal rectangles of area sizes proportional to frequencies;
- \code{"vr"} ... boxes as vertical rectangles of area sizes proportional to frequencies.}

\item{round}{A Boolean option that determines whether frequencies are
rounded to the nearest integer. Default: \code{round = TRUE}.}

\item{title.lbl}{Text label to set plot title.}

\item{col.boxes}{Colors of boxes (a single color or vector of 7 colors).}
}
\value{
Nothing (NULL).
}
\description{
\code{plot_tree} computes and visualizes a tree diagram of
natural frequencies (typically rounded integers) given basic
probabilities -- \code{\link{prev}} and \code{\link{sens}},
and \code{\link{spec}} or \code{\link{fart}} (\code{spec = 1 - fart})
-- and a population size of \code{\link{N}} individuals.
}
\details{
The option \code{area} (as 2 characters) allows specifying
4 different box shapes and sizes:

\enumerate{
 \item \code{"no"} shows all boxes in the same size (default);
 \item \code{"sq"} shows boxes as squares with area sizes proportional to frequencies;
 \item \code{"hr"} shows boxes as horizontal rectangles of area sizes proportional to frequencies;
 \item \code{"vr"} shows boxes as vertical rectangles of area sizes proportional to frequencies.
The resulting shapes and their relative proportions correspond to the areas in \code{\link{plot_mosaic}}.
}

If a prevalence value \code{\link{prev}} is provided, a new list of
natural frequencies \code{\link{freq}} is computed by \code{\link{comp_freq}}.
By contrast, if no prevalence value \code{\link{prev}} is provided,
the values currently contained in \code{\link{freq}} are used.
By default, \code{\link{comp_freq}} rounds frequencies to nearest integers
to avoid decimal values in \code{\link{freq}}.

\code{plot_tree} requires and uses the R package "diagram"
(\code{library("diagram")}).
}
\examples{
# Plotting existing freq:
plot_tree()
plot_tree(title.lbl = "")
plot_tree(N = 33)
plot_tree(N = NA)

## Computing and plotting new frequencies:
plot_tree(prev = 1/3)
plot_tree(prev = 1/3, N = 55)
plot_tree(prev = 1/3, N = NA)
plot_tree(prev = 1/3, round = FALSE)
plot_tree(prev = .10, sens = .90, spec = 1/3, N = 100)
plot_tree(prev = .10, sens = .90, spec = NA, fart = 1/3, N = 33)
plot_tree(prev = .10, sens = .90, spec = 1/3, fart = NA, N = NA)
plot_tree(prev = .10, sens = .90, spec = NA, fart = 1/3, N = NA)

## Area trees:
plot_tree(area = "sq")
plot_tree(area = "sq", round = FALSE)
plot_tree(area = "hr")
plot_tree(area = "vr", round = FALSE)

## Custom colors and shadows:
plot_tree(prev = .08, sens = .92, spec = .95, N = 10000, area = "hr")
plot_tree(area = "sq", col.boxes = "gold", col.border = "steelblue4", col.shadow = "steelblue4", cex.shadow = .008)
plot_tree(N = NA, area = "vr", col.txt = "steelblue4", col.boxes = "lightyellow", col.border = grey(.3, .7), cex.shadow = .008, col.shadow = grey(.1, .9))

}
\seealso{
\code{\link{num}} for basic numeric parameters;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes frequencies from probabilities;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{pal}} for current color settings;
\code{\link{txt}} for current text settings;
\code{\link{comp_min_N}} computes a suitable population size \code{\link{N}} (if missing)
}
