% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\docType{data}
\name{prev}
\alias{prev}
\title{The prevalence (baseline probability) of a condition.}
\format{An object of class \code{numeric} of length 1.}
\usage{
prev
}
\description{
\code{prev} defines a condition's prevalence value:
The baseline probability of the condition being \code{TRUE}.
}
\details{
Understanding or obtaining the prevalence value \code{prev}:

\itemize{

\item \code{prev} is the (non-conditional) probability:

\code{prev = p(condition = TRUE)}

\item In natural frequencies,
  \code{prev} is the ratio of individuals for which
  \code{condition = TRUE} divided by the
  number \code{\link{N}} of all individuals
  in the population:

\code{prev = n(condition = TRUE) / \link{N}}

\item \code{prev} is a feature of the population
  and condition, but independent of the decision process
  or diagnostic procedure.

\item The value of \code{prev} does not depend
  on features of the decision process or diagnostic procedure.
  However, \code{prev} must be taken into account when
  computing the conditional probabilities
  \code{\link{sens}}, \code{\link{spec}}, \code{\link{fart}},
  \code{\link{ppv}}, and \code{\link{npv}}
  (as derived parameters).

}
}
\examples{
prev <- .10     # => sets a prevalence value of 10\%
prev <- 10/100  # => (condition = TRUE) for 10 out of 100 individuals
is_prob(prev)   # => TRUE (as prev is a probability)

}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities

Other basic parameters: \code{\link{N}},
  \code{\link{fart}}, \code{\link{sens}},
  \code{\link{spec}}
}
\keyword{datasets}
