% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{print.box}
\alias{print.box}
\title{Print a frequency box object}
\usage{
\method{print}{box}(x)
}
\arguments{
\item{x}{A box object}
}
\description{
\code{print.box} is a utility method that prints a box object.
}
\seealso{
Other utility functions: 
\code{\link{as_pb}()},
\code{\link{as_pc}()},
\code{\link{plot.box}()}
}
\concept{utility functions}
