% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskyr_class.R
\name{print.table.riskyr}
\alias{print.table.riskyr}
\title{Printing tabular risk information.}
\format{Printed output of a "table.riskyr" object.}
\usage{
\method{print}{table.riskyr}(x, space = 2)
}
\arguments{
\item{x}{An object of class "table.riskyr", usually a result of a call to \code{table.riskyr}.}

\item{space}{How much space should be included left and right of the table entries?}
}
\description{
\code{print.table.riskyr} provides a \code{print} method for objects of class "table.riskyr".
}
\examples{
table(scenarios$n4)

}
\seealso{
Other print functions: \code{\link{print.summary.riskyr}}
}
\concept{print functions}
