% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\docType{data}
\name{sens}
\alias{sens}
\title{The sensitivity of a decision process or diagnostic procedure.}
\format{An object of class \code{numeric} of length 1.}
\usage{
sens
}
\description{
\code{sens} defines a decision's sensitivity value:
The conditional probability of the decision being positive
if the condition is \code{TRUE}.
}
\details{
Understanding or obtaining the sensitivity value \code{sens}:

\itemize{

\item \code{sens} is the conditional probability
  for a (correct) positive decision given that
  the condition is \code{TRUE}:

\code{sens = p(decision = positive | condition = TRUE)}

\code{sens} is the opposite conditional probability
  -- but not the complement --
  of the positive predictive value \code{\link{ppv}}:

\code{ppv = p(condition = TRUE | decision = positive)}

\item In natural frequencies,
  \code{sens} is the ratio of individuals for which
  \code{decision = positive} divided by the number of
  individuals for which \code{condition = TRUE}:

\code{sens = n(decision = positive) / n(condition = TRUE)}

\item \code{sens} is a feature of a decision process
  or diagnostic procedure and a measure of
  correct decisions (true positives).

However, due to being a conditional probability,
  the value of \code{sens} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
sens <- .85     # => sets a sensitivity value of 85\%
sens <- 85/100  # => (decision = positive) for 85 people out of 100 people for which (condition = TRUE)
is_prob(sens)   # => TRUE (as sens is a probability)

}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities

Other basic parameters: \code{\link{N}},
  \code{\link{fart}}, \code{\link{prev}},
  \code{\link{spec}}
}
\keyword{datasets}
