% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{sim_check_1}
\alias{sim_check_1}
\title{Simulation to verify base-to-decimal conversion functions}
\usage{
sim_check_1(N = 100, fb = FALSE)
}
\description{
\code{sim_check_1} wraps a simulation to verify that 3 conversion functions
\code{base_dec()}, \code{recursive base_2_dec()}, and \code{ds4psy::base2dec()} yield identical results.
}
