% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_util.R
\name{sim_check_2}
\alias{sim_check_2}
\title{Simulation to verify decimal-to-base and base-to-decimal conversion}
\usage{
sim_check_2(N = 100)
}
\description{
\code{sim_check_2} wraps a simulation to verify that combining the 2 numeral conversion functions
recursive \code{dec_2_base()} and recursive \code{base_2_dec()} yield the original decimal number.
}
