% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mlm.R
\name{trans}
\alias{trans}
\title{Transform a 2x2 matrix (into a table of probabilities/conditional probabilities).}
\usage{
trans(mx, margin = 0, as_pc = FALSE, n_digits = 3)
}
\arguments{
\item{mx}{A 2x2 matrix (as numeric contingency table, required).}

\item{margin}{Margin to conditionalize table values (numeric, from 0 to 3).
Default: \code{margin = 0} (i.e., unconditional probabilities).}

\item{as_pc}{Boolean: Convert probabilities into percentages?
Default: \code{as_pc = FALSE}.}

\item{n_digits}{Number of decimal places to which result is rounded.
Default: \code{n_digits = 3}.}
}
\value{
A numeric 2x2 table.
}
\description{
\code{trans} converts a 2x2 matrix \code{mx}
(i.e., a numeric contingency table of frequency counts)
into a table of probabilities or marginal probabilities.
}
\details{
By default, \code{trans} converts a 2x2 matrix of frequency counts
into the corresponding probability values (\code{margin = 0}).

Setting \code{margin} to 1 to 3 computes
conditional probabilities in a
by row (\code{margin = 1}),
by column (\code{margin = 2}), or
by diagonal (\code{margin = 3}) fashion.

The tables for \code{margin = 0} to \code{margin = 2}
are computed by the \strong{base} R function \code{\link{prop.table}}.
}
\examples{
# Mammography problem:
abcd <- c(8, 95, 2, 895)  # Frequencies (Gigerenzer & Hoffrage, 1995)
mp <- frame(data = abcd, x = "Condition", y = "Test",
            x_levels = c("cancer", "no cancer"),
            y_levels = c("positive", "negative"))

trans(mp)
trans(mp, margin = 0)  # by-cell: probabilities
trans(mp, margin = 1)  # by-row: conditional prob
trans(mp, margin = 2)  # by-col
trans(mp, margin = 3)  # by-diagonals

trans(mp, as_pc = TRUE, n_digits = 2)  # as percentages

# The following must sum to 1:
sum(trans(mp, margin = 0))      # 4 cell values
rowSums(trans(mp, margin = 1))  # 2 row sums
colSums(trans(mp, margin = 2))  # 2 col sums

}
\seealso{
\code{\link{frame}} for creating a 2x2 matrix;
\code{\link{is_matrix}} for verifying a 2x2 matrix.

Other matrix lens model functions: 
\code{\link{frame}()}
}
\concept{matrix lens model functions}
