% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_prob_freq.R
\name{trans_prob}
\alias{trans_prob}
\title{Translate probabilities into frequencies.}
\usage{
trans_prob(prev = prob$prev, sens = prob$sens, spec = prob$spec,
  N = freq$N, round = TRUE)
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).}

\item{N}{The number of individuals in the population.
If \code{\link{N}} is unknown (\code{NA}),
a suitable minimum value is computed by \code{\link{comp_min_N}}.}

\item{round}{A Boolean value that determines whether frequencies are
rounded to the nearest integer. Default: \code{round = TRUE}.}
}
\value{
A list \code{\link{freq}} containing 9 frequency values.
}
\description{
\code{trans_prob} translates information provided by
3 essential probabilities --
\code{\link{prev}}, \code{\link{sens}}, and \code{\link{spec}}
-- into frequencies for a population of \code{\link{N}} individuals.
It returns a list of 9 frequencies \code{\link{freq}}
as its output.
}
\details{
\code{trans_prob} is a wrapper function for \code{\link{comp_freq}},
which is the frequency counterpart to the probability
function \code{\link{comp_prob}}.

By default, the values of \code{\link{prev}}, \code{\link{sens}},
and \code{\link{spec}} are initialized to the probability information
currently contained in \code{\link{prob}}.

The value of \code{\link{N}} uses the frequency information
currently contained in \code{\link{freq}} as its default.
If \code{\link{N}} is unknown (\code{NA}),
a suitable minimum value can be computed by \code{\link{comp_min_N}}.

In addition to \code{\link{prev}}, both
\code{\link{sens}} and \code{\link{spec}} are necessary arguments.
If only their complements \code{\link{mirt}} or \code{\link{fart}}
are known, first use \code{\link{comp_prob_comp}},
\code{\link{comp_comp_pair}}, or \code{\link{comp_complete_prob_set}}
to obtain the 3 essential probabilities.

By default, \code{trans_prob} and \code{\link{comp_freq}}
round frequencies to nearest integers to avoid decimal values in
\code{\link{freq}}. Use \code{round = FALSE}
to switch off rounding.

Key relationships:

\enumerate{

\item to probabilities:
A population of \code{\link{N}} individuals can be split into 2 subsets
in 2 different ways:

\enumerate{
  \item by condition:
  The frequency \code{\link{cond.true}} depends on the prevalence \code{\link{prev}}
  and
  the frequency \code{\link{cond.false}} depends on the prevalence's complement \code{1 - \link{prev}}.

\item by decision:
  The frequency \code{\link{dec.pos}} depends on the proportion of positive decisions \code{\link{ppod}}
  and
  the frequency \code{\link{dec.neg}} depends on the proportion of negative decisions \code{1 - \link{ppod}}.

}

The population size \code{\link{N}} is a free parameter (independent of the
essential probabilities \code{\link{prev}}, \code{\link{sens}}, and \code{\link{spec}}).

If \code{\link{N}} is unknown (\code{NA}), a suitable minimum value can be computed by \code{\link{comp_min_N}}.

\item to other frequencies:
  In a population of size \code{\link{N}}
  the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}} (by condition)

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}} (by decision)

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}} (by condition x decision)
  }

}
}
\examples{
trans_prob()          # => ok, using probability info currently contained in prob
length(trans_prob())  # => a list containing 9 frequencies
freq <- trans_prob()  # => translates and stores list in freq

# Ways to succeed:
trans_prob(prev = 1, sens = 1, spec = 1, 101)  # => ok + warning: N hits (TP)
trans_prob(prev = 1, sens = 1, spec = 0, 102)  # => ok + warning: N hits
trans_prob(prev = 1, sens = 0, spec = 1, 103)  # => ok + warning: N misses (FN)
trans_prob(prev = 1, sens = 0, spec = 0, 104)  # => ok + warning: N misses
trans_prob(prev = 0, sens = 1, spec = 1, 105)  # => ok + warning: N correct rejections (TN)
trans_prob(prev = 0, sens = 1, spec = 0, 106)  # => ok + warning: N false alarms (FP)

# Watch out for:
trans_prob(prev = 1, sens = 1, spec = 1, N = NA)  # => ok + additional warning that N = 1 was computed
trans_prob(prev = 1, sens = 1, spec = 1, N =  0)  # => ok, but all 0 + warning (extreme case: N hits, NPV = NaN)
trans_prob(prev = .5, sens = .5, spec = .5, N = 10, round = TRUE)   # => ok, but all rounded (increasing errors: mi and fa)
trans_prob(prev = .5, sens = .5, spec = .5, N = 10, round = FALSE)  # => ok, but not rounded

# Ways to fail:
trans_prob(prev = NA,  sens = 1, spec = 1,  100)   # => NAs + no warning (prev NA)
trans_prob(prev = 1,  sens = NA, spec = 1,  100)   # => NAs + no warning (sens NA)
trans_prob(prev = 1,  sens = 1,  spec = NA, 100)   # => NAs + no warning (spec NA)
trans_prob(prev = 8,  sens = 1,  spec = 1,  100)   # => NAs + warning (prev beyond range)
trans_prob(prev = 1,  sens = 8,  spec = 1,  100)   # => NAs + warning (sens & spec beyond range)


}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{freq}} contains current frequency information;
\code{\link{init_freq}} initializes frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{comp_prob_comp}} computes a probability's complement;
\code{\link{comp_comp_pair}} computes pairs of complements;
\code{\link{comp_complete_prob_set}} completes valid sets of probabilities;
\code{\link{comp_min_N}} computes a suitable population size \code{\link{N}} (if missing).

Other functions computing frequencies: \code{\link{comp_freq_freq}},
  \code{\link{comp_freq}}, \code{\link{comp_min_N}}
}
