% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_txt.R
\docType{data}
\name{txt_alt}
\alias{txt_alt}
\title{Alternative text labels (TP, FN, FP, TN).}
\format{An object of class \code{list} of length 20.}
\usage{
txt_alt
}
\description{
\code{txt_alt} is initialized to alternative text labels
to define a frequency naming scheme in which
(hi, mi, fa, cr) are called (TP, FN, FP, TN).
}
\details{
See \code{\link{txt}} for default text information.

Assign \code{txt <- txt_alt} to use as default text labels
throughout the \code{riskyr} package.
}
\examples{
txt_alt       # shows all text labels
txt_alt["hi"]  # shows the current color for hits ("TP")
txt_alt["hi"] <- "hit" # defines a new label for hits (true positives, TP)

}
\seealso{
\code{\link{txt}} contains current color information;
\code{\link{init_txt}} initializes color information.

Other lists containing current scenario information: \code{\link{accu}},
  \code{\link{freq}}, \code{\link{num}},
  \code{\link{pal_4c}}, \code{\link{pal_bw}},
  \code{\link{pal}}, \code{\link{prob}}, \code{\link{txt}}
}
\keyword{datasets}
