% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_txt.R
\docType{data}
\name{txt_org}
\alias{txt_org}
\title{List of original values of text elements.}
\format{An object of class \code{list} of length 21.}
\usage{
txt_org
}
\description{
\code{txt_org} is a copy of the initial list of text elements
to define all scenario titles and labels.
}
\details{
See \code{\link{txt}} for details and default text information.

Assign \code{txt <- txt_org} to re-set default text labels.
}
\examples{
txt_org        # shows original text labels
txt_org["hi"]  # shows the original label for hits ("hi")
txt_org["hi"] <- "TP" # defines a new label for hits (true positives, TP)

}
\seealso{
\code{\link{txt}} contains current text information;
\code{\link{init_txt}} initializes text information;
\code{\link{pal}} contains current color information;
\code{\link{init_pal}} initializes color information.

Other lists containing current scenario information: \code{\link{accu}},
  \code{\link{freq}}, \code{\link{num}},
  \code{\link{pal_bw}}, \code{\link{pal_kn}},
  \code{\link{pal_mod}}, \code{\link{pal_org}},
  \code{\link{pal_rgb}}, \code{\link{pal_vir}},
  \code{\link{pal}}, \code{\link{prob}},
  \code{\link{txt_TF}}, \code{\link{txt}}
}
\concept{lists containing current scenario information}
\keyword{datasets}
