% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_accu.R
\docType{data}
\name{accu}
\alias{accu}
\title{A list containing current accuracy information.}
\format{An object of class \code{list} of length 5.}
\usage{
accu
}
\description{
\code{accu} contains current accuracy information
returned by the corresponding generating function
\code{\link{comp_accu}}.
}
\examples{
accu <- comp_accu()  # => computes current accuracy information and stores result in accu


}
\seealso{
The corresponding generating function \code{\link{comp_accu}};
\code{\link{num}} for basic numeric parameters;
\code{\link{freq}} for current frequency information;
\code{\link{prob}} for current probability information;
\code{\link{txt}} for current text settings.

Other lists containing current scenario information: \code{\link{freq}},
  \code{\link{num}}, \code{\link{pal}}, \code{\link{prob}},
  \code{\link{txt}}
}
\keyword{datasets}
