% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\name{comp_complement}
\alias{comp_complement}
\title{Compute a probability's complement (if missing).}
\usage{
comp_complement(spec, fart)
}
\arguments{
\item{spec}{A specificity value (i.e., the conditional probability
of a negative decision provided that the condition is FALSE).
\code{spec} is optional when is complement \code{fart} is provided.}

\item{fart}{A false alarm rate (i.e., the conditional probability
of a positive decision provided that the condition is FALSE).
\code{fart} is optional when its complement \code{spec} is provided.}
}
\value{
A vector \code{v} containing two scalars \code{c(spec, fart)}
with \code{spec <- v[1]} and \code{fart <- v[2]}).
}
\description{
\code{comp_complement} is a function that takes one or two probabilities
--  a specificity \code{spec} and a false alarm rate \code{fart} --
as inputs. If either of them is missing (\code{NA}), it computes the complement
of the other one and returns both probabilities.
}
\details{
This function does nothing when both arguments are provided
(i.e., \code{!is.na(spec) & !is.na(fart)}) and only issues
a warning if both arguments are missing
(i.e., \code{is.na(spec) & is.na(fart)}).
Use \code{\link{is_complement}} to verify that
two provided values actually are complements.
}
\examples{
comp_complement(1, 0)   # => 1 0
comp_complement(0, 1)   # => 0 1
comp_complement(1, NA)  # => 1 0
comp_complement(NA, 1)  # => 0 1

comp_complement(NA, NA) # => NA NA + warning
comp_complement(1, 1)   # => 1 1 + NO warning (as is_complement is not applied here)
comp_complement(8, 8)   # => 8 8 + NO warning (as is_prob or is_valid are not applied here)

}
\seealso{
\code{\link{is_complement}} to verify numeric complements;
\code{\link{is_valid}} to verify valid quadruples of probabilities

Other conversion functions: \code{\link{comp_fart}},
  \code{\link{comp_spec}}
}
