% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_popu.R
\name{comp_popu}
\alias{comp_popu}
\title{Compute a population table from frequencies.}
\usage{
comp_popu(hi = freq$hi, mi = freq$mi, fa = freq$fa, cr = freq$cr,
  cond.true.lbl = txt$cond.true.lbl, cond.false.lbl = txt$cond.false.lbl,
  dec.pos.lbl = txt$dec.pos.lbl, dec.neg.lbl = txt$dec.neg.lbl,
  sdt.hi.lbl = txt$sdt.hi.lbl, sdt.mi.lbl = txt$sdt.mi.lbl,
  sdt.fa.lbl = txt$sdt.fa.lbl, sdt.cr.lbl = txt$sdt.cr.lbl)
}
\value{
A data frame \code{popu}
containing \code{\link{N}} rows (individual cases)
and 3 columns (\code{"Truth", "Decision", "SDT"})
encoded as ordered factors
(with 2, 2, and 4 levels, respectively).
}
\description{
\code{comp_popu} is a function that computes
a table \code{\link{popu}} (as an R data frame)
from the current frequency information
(contained in \code{\link{freq}}).
}
\details{
\code{comp_popu} also uses the current text settings
contained in \code{\link{txt}}.

A visualization of the current population
contained in \code{\link{popu}}
is provided by \code{\link{plot_iconarray}}.
}
\examples{
comp_popu(hi = 4, mi = 1, fa = 2, cr = 3)  # => computes a table of N = 10 cases

popu <- comp_popu()  # => initializes popu (with current values of freq and txt)
dim(popu)            # => N x 3
head(popu)           # => shows head of data frame


}
\seealso{
The corresponding data frame \code{\link{popu}};
\code{\link{num}} for basic numeric parameters;
\code{\link{freq}} for current frequency information;
\code{\link{txt}} for current text settings;
\code{\link{pal}} for current color settings
}
