% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num_prob.R
\name{comp_prob}
\alias{comp_prob}
\title{Compute derived probabilities from basic probabilities.}
\usage{
comp_prob(prev = num$prev, sens = num$sens, mirt = NA, spec = num$spec,
  fart = NA)
}
\arguments{
\item{prev}{The condition's prevalence value \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity value \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{\link{mirt}} is provided.}

\item{mirt}{The decision's miss rate value \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{\link{sens}} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{\link{fart}} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{\link{spec}} is provided.}
}
\value{
A list \code{\link{prob}} containing 8 probability values.
}
\description{
\code{comp_prob} is a function that computes derived probabilities
(typically conditional probabilities) from basic probabilities --
\code{\link{prev}} and \code{\link{sens}}, and
\code{\link{spec}} or \code{\link{fart}} (\code{spec = 1 - fart}).
}
\details{
By default, \code{comp_prob} assumes that sufficient
basic probabilities (e.g., \code{\link{prev}}, \code{\link{sens}},
and either \code{\link{spec}} or \code{\link{fart}}) are provided
as inputs and then computes and returns derived probabilities (e.g., the
predictive values \code{\link{PPV}} and \code{\link{NPV}}, as well
as their complements \code{\link{FDR}} and \code{\link{FOR}})
as its output (a list of probabilities \code{\link{prob}}).

\code{comp_prob} is the probability counterpart to the
frequency function \code{\link{comp_freq}}.

Extreme probabilities (sets containing two or more
probabilities of 0 or 1) may yield unexpected values
(e.g., predictive values \code{\link{PPV}} or \code{\link{NPV}}
turning \code{NaN} when \code{\link{is_extreme}}
evaluates to \code{TRUE}).
}
\examples{
comp_prob()             # => ok, using current defaults
length(comp_prob())     # => 8

# Ways to succeed:
comp_prob(.999, 1, 1)   # => ok
comp_prob(1, .999, 1)   # => ok

# Watch out for extreme cases:
comp_prob(1, 0, 1)      # => ok, but with warnings (as PPV & FDR are NaN)
comp_prob(1, 0, 0)      # => ok, but with warnings (as PPV & FDR are NaN)
comp_prob(1, 0, NA, 0)  # => ok, but with warnings (as PPV & FDR are NaN)
comp_prob(1, 0, NA, 1)  # => ok, but with warnings (as PPV & FDR are NaN)

comp_prob(1, 1, 0)      # => ok, but with warnings (as NPV & FOR are NaN)
comp_prob(1, 1, 1)      # => ok, but with warnings (as NPV & FOR are NaN)
comp_prob(1, 1, 1, NA)  # => ok, but with warnings (as NPV & FOR are NaN)
comp_prob(1, 1, NA, 1)  # => ok, but with warnings (as NPV & FOR are NaN)

# Ways to fail:
comp_prob(NA, 1, 1, NA)  # => NAs + warning: prev not numeric
comp_prob(1, NA, 1, NA)  # => NAs + warning: sens not numeric
comp_prob(8,  1, 1, NA)  # => NAs + warning: prev no probability
comp_prob(1,  8, 1, NA)  # => NAs + warning: sens no probability
comp_prob(1,  1, 1,  1)  # => NAs and warning: is_complement not in tolerated range

}
\seealso{
\code{\link{prob}} contains current probability information;
\code{\link{num}} contains basic parameter values;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{pal}} contains current color settings;
\code{\link{txt}} contains current text settings;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes frequencies from probabilities;
\code{\link{is_valid_prob_set}} verifies sets of probability inputs;
\code{\link{is_extreme_prob_set}} verifies sets of extreme probabilities;
\code{\link{comp_min_N}} computes a suitable minimum population size \code{\link{N}}.

Other functions computing probabilities: \code{\link{comp_FDR}},
  \code{\link{comp_FOR}}, \code{\link{comp_NPV}},
  \code{\link{comp_PPV}}, \code{\link{comp_comp_pair}},
  \code{\link{comp_complete_prob_set}},
  \code{\link{comp_fart}}, \code{\link{comp_mirt}},
  \code{\link{comp_prob_comp}},
  \code{\link{comp_prob_freq}}, \code{\link{comp_sens}},
  \code{\link{comp_spec}}
}
