% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_prob_prob.R
\name{comp_spec}
\alias{comp_spec}
\title{Compute a decision's specificity from its false alarm rate.}
\usage{
comp_spec(fart)
}
\arguments{
\item{fart}{The decision's false alarm rate \code{\link{fart}} as a probability.}
}
\value{
The decision's specificity \code{\link{spec}} as a probability.
}
\description{
\code{comp_spec} is a conversion function that takes a false alarm rate \code{\link{fart}}
-- given as a probability (i.e., a numeric value in the range from 0 to 1) --
as its input, and returns the corresponding specificity \code{\link{spec}}
-- also as a probability -- as its output.
}
\details{
The specificity \code{\link{spec}} and the false alarm rate \code{\link{fart}}
are complements (\code{spec = (1 - fart)}) and both features of
the decision process (e.g., a diagnostic test).

The function \code{comp_spec} is complementary to the conversion function
\code{\link{comp_fart}} and uses the generic function
\code{\link{comp_complement}}.
}
\examples{
comp_spec(2)                      # => NA + warning (beyond range)
comp_spec(1/3)                    # => 0.6666667
comp_spec(comp_complement(0.123))  # => 0.123

}
\seealso{
\code{\link{comp_complement}} computes a probability's complement;
\code{\link{is_complement}} verifies probability complements;
\code{\link{comp_prob}} computes current probability information;
\code{\link{prob}} contains current probability information;
\code{\link{is_prob}} verifies probabilities.

Other functions computing probabilities: \code{\link{comp_FDR}},
  \code{\link{comp_FOR}}, \code{\link{comp_NPV}},
  \code{\link{comp_PPV}}, \code{\link{comp_accu}},
  \code{\link{comp_comp_pair}},
  \code{\link{comp_complement}},
  \code{\link{comp_complete_prob_set}},
  \code{\link{comp_fart}}, \code{\link{comp_mirt}},
  \code{\link{comp_ppod}}, \code{\link{comp_prob_freq}},
  \code{\link{comp_prob}}, \code{\link{comp_sens}}
}
