% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\name{comp_spec}
\alias{comp_spec}
\title{Compute a decision's specificity from its false alarm rate.}
\usage{
comp_spec(fart)
}
\arguments{
\item{fart}{A false alarm rate given as a probability
(i.e., a numeric value in the range from 0 to 1).}
}
\value{
A specificity as a probability
(i.e., a numeric value in the range from 0 to 1).
}
\description{
\code{comp_spec} is a function that takes a false alarm rate \code{fart}
-- given as a probability (i.e., a numeric value in the range from 0 to 1) --
as its input, and returns the corresponding specificity \code{spec}
-- also as a probability -- as its output.
}
\details{
False alarm rate and specificity are both features of the decision process
(e.g., a diagnostic test).
The specificity and false alarm rate are complements (i.e., \code{spec = 1 - fart}).

\code{comp_spec} is complementary to the conversion function
\code{\link{comp_fart}}.
}
\examples{
comp_spec(2)                 # => NA + Warning that 2 is not in 0 to 1 range
comp_spec(2/3)               # => 0.3333333
comp_spec(comp_fart(0.123))  # => 0.123
}
\seealso{
\code{\link{is_complement}} to verify numeric complements;
\code{\link{as_pc}} to display a probability as a percentage

Other conversion functions: \code{\link{comp_complement}},
  \code{\link{comp_fart}}
}
