% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prob.R
\docType{data}
\name{fart}
\alias{fart}
\alias{fallout}
\alias{FPR}
\title{The false alarm rate (or false positive rate) of a decision process or diagnostic procedure.}
\format{An object of class \code{numeric} of length 1.}
\usage{
fart
}
\description{
\code{fart} defines a decision's false alarm rate
(or the rate of false positives): The conditional probability
of the decision being positive if the condition is FALSE.
}
\details{
Understanding or obtaining the false alarm rate \code{fart}:

\itemize{

\item Definition:
  \code{fart} is the conditional probability
  for an incorrect positive decision given that
  the condition is \code{FALSE}:

\code{fart = p(decision = positive | condition = FALSE)}

or the probability of a false alarm.

\item Alternative names:
  \code{fallout}, false positive rate (\code{FPR}),
  rate of type-I errors

\item Relationships:

a. \code{fart} is the complement of the
  specificity \code{\link{spec}}:

\code{fart = 1 - spec}

b. \code{fart} is the opposite conditional probability
  -- but not the complement --
  of the false discovery rate
  or false detection rate \code{\link{FDR}}:

\code{FDR = p(condition = FALSE | decision = negative)}

\item In natural frequencies,
  \code{fart} is the ratio of individuals for which
  \code{decision = positive} divided by the number of
  individuals for which \code{condition = FALSE}:

\code{fart = n(decision = positive) / n(condition = FALSE)}

\item \code{fart} is a feature of a decision process
  or diagnostic procedure and a measure of
  incorrect decisions (false positives).

However, due to being a conditional probability,
  the value of \code{fart} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
fart <- .25     # => sets a false alarm rate of 25\%
fart <- 25/100  # => (decision = positive) for 25 people out of 100 people for which (condition = FALSE)
is_prob(fart)   # => TRUE (as fart is a probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{Wikipedia} for additional information.
}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities;
\code{\link{comp_fart}} computes \code{fart}
as the complement of \code{\link{spec}}

Other basic parameters: \code{\link{N}},
  \code{\link{mirt}}, \code{\link{prev}},
  \code{\link{sens}}, \code{\link{spec}}
}
\keyword{datasets}
