% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_freq.R
\docType{data}
\name{freq}
\alias{freq}
\title{List current frequency information.}
\format{An object of class \code{list} of length 9.}
\usage{
freq
}
\description{
\code{freq} is a list of named numeric variables
containing 9 (natural) frequencies:
\enumerate{
 \item the population size \code{\link{N}}
 \item the number of cases for which \code{cond.true}
 \item the number of cases for which \code{cond.false}
 \item the number of cases for which \code{dec.pos}
 \item the number of cases for which \code{dec.neg}
 \item the number true positives, or hits (\code{hi})
 \item the number false negatives, or misses (\code{mi})
 \item the number false positives, or false alarms (\code{fa})
 \item the number true negatives, or correct rejections (\code{cr})
}
}
\details{
These frequencies are computed from basic probabilities
(contained in \code{\link{num}}) and computed by using
\code{\link{comp_freq}}.

The list \code{freq} is the frequency counterpart
to the list containing probability information \code{\link{prob}}.

Natural frequencies are always expressed in
relation to the current population of
size \code{\link{N}}.

Visualizations of the current frequency information
are provided by \code{\link{plot_tree}} and
\code{\link{plot_mosaic}}.
}
\examples{
freq <- comp_freq()  # => initialize freq to default parameters
freq                 # => show current values
length(freq)         # => 9

}
\seealso{
\code{\link{comp_freq}} computes current frequency information;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{prob}} contains current probability information

Other lists containing scenario settings: \code{\link{num}},
  \code{\link{prob}}
}
\keyword{datasets}
