% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_freq.R
\docType{data}
\name{N}
\alias{N}
\title{Number of individuals in the population.}
\format{An object of class \code{numeric} of length 1.}
\usage{
N
}
\description{
\code{N} is a frequency that describes the
number of individuals in the current population
(i.e., the overall number of cases considered).
}
\details{
\emph{Relationships}:
In a population of size \code{\link{N}}
the following relationships hold:

\itemize{

\item \code{\link{N} = \link{cond.true} + \link{cond.false}}

\item \code{\link{N} = \link{dec.pos} + \link{dec.neg}}

\item \code{\link{N} = \link{hi} + \link{mi} + \link{fa} + \link{cr}}
}

The current frequency information is computed by
\code{\link{comp_freq}} and contained in a list
\code{\link{freq}}.

If \code{N} is unknown, a suitable minimum value
can be computed by \code{\link{comp_min_N}}.
}
\examples{
N <- 1000   # => sets a population size of 1000
is_freq(N)  # => TRUE
is_prob(N)  # => FALSE (as N is no probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Statistical_population}{Wikipedia} for additional information.
}
\seealso{
\code{\link{is_freq}} verifies frequencies;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{freq}} contains current frequency information;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information.

Other frequencies: \code{\link{cr}}, \code{\link{fa}},
  \code{\link{hi}}, \code{\link{mi}}
}
\keyword{datasets}
