% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\docType{data}
\name{N}
\alias{N}
\title{The number of individuals in the population.}
\format{An object of class \code{numeric} of length 1.}
\usage{
N
}
\description{
\code{N} defines how many individuals make up
the current population (i.e., the overall
number of cases considered).
}
\details{
A frequency (integer) value.

The following relationships hold for natural
frequencies in a population:

\itemize{

\item N = n(condition true) + n(condition false)

\item n(positive decisions) + n(negative decisions)

\item N = n(hi) + n(mi) + n(fa) + n(cr)
}

The current frequency information is computed by
\code{\link{comp_freq}} and contained in a list
\code{\link{freq}}.

If \code{N} is unknown, a suitable minimum value
can be computed by \code{\link{comp_min_N}}.
}
\examples{
N <- 1000   # => sets a population size of 1000
is_freq(N)  # => TRUE
is_prob(N)  # => FALSE (as N is no probability)

}
\seealso{
\code{\link{comp_min_N}} computes a suitable
minimum value of \code{N} for given probabilities;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{freq}} contains basic frequency variables;
\code{\link{comp_freq}} computes natural frequencies from probabilities;
\code{\link{comp_prob}} computes derived probabilities

Other basic parameters: \code{\link{fart}},
  \code{\link{prev}}, \code{\link{sens}},
  \code{\link{spec}}
}
\keyword{datasets}
