% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num.R
\docType{data}
\name{num}
\alias{num}
\title{List current values of basic numeric variables.}
\format{An object of class \code{list} of length 5.}
\usage{
num
}
\description{
\code{num} is initialized to a list of named numeric variables containing
four basic probabilities (\code{prev}, \code{sens}, \code{spec}, and \code{fart})
and one frequency parameter (the population size \code{N}).
}
\seealso{
\code{\link{init_num}} to initialize basic parameter values

Other lists containing basic scenario settings: \code{\link{pal}},
  \code{\link{txt}}
}
\keyword{datasets}
