% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{plot.box}
\alias{plot.box}
\title{Plot information of a riskyr object.}
\usage{
\method{plot}{box}(x, cur_freq = freq, cur_txt = txt, cur_pal = pal,
  ...)
}
\arguments{
\item{cur_freq}{The frequency information related to this box object.}

\item{cur_txt}{The text information associate with this box object.}

\item{cur_pal}{The color information related to this box object.}

\item{...}{Additional (graphical) parameters to be passed to the
underlying plotting functions.}

\item{obj}{The object of class "box" to be plotted.}
}
\description{
\code{plot.box} is a utility method that allows to plot low level boxes for \code{riskyr} plots.
}
\details{
\code{plot.riskyr} also uses the text settings
specified in the "riskyr" object.
}
\seealso{
Other utility functions: \code{\link{as_pb}},
  \code{\link{as_pc}}
}
\concept{utility functions}
