% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot bar charts of population frequencies.}
\usage{
plot_bar(prev = num$prev, sens = num$sens, mirt = NA, spec = num$spec,
  fart = NA, N = num$N, by = "all", show.freq = FALSE,
  show.prob = FALSE, show.accu = TRUE, w.acc = 0.5,
  title.lbl = txt$scen.lbl, col.pal = pal)
}
\arguments{
\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{mirt}{The decision's miss rate \code{\link{mirt}}
(i.e., the conditional probability of a negative decision
provided that the condition is \code{TRUE}).
\code{mirt} is optional when its complement \code{sens} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the population.
(This value is not represented in the plot,
but used when new frequency information \code{\link{freq}}
and a new population table \code{\link{popu}}
are computed from scratch from current probabilities.)}

\item{by}{A character code specifying the perspective
(or 1st category by which the population is split into subsets)
with 2 options:
  \enumerate{
  \item \code{"cd"} ... by condition;
  \item \code{"dc"} ... by decision;
  \item \code{"all"} ... all different subdivisions.
  }}

\item{show.freq}{Boolean option for showing frequencies
(i.e., of \code{\link{hi}}, \code{\link{mi}}, \code{\link{fa}}, and
\code{\link{cr}}) in the plot.
Default: \code{show.freq = FALSE}.}

\item{show.prob}{Boolean option for showing visual help lines to mark
generating metrics (e.g., \code{\link{prev}}, \code{\link{sens}}, and
\code{\link{spec}}) in the plot.
Default: \code{show.prob = FALSE}.}

\item{show.accu}{Option for showing current
accuracy metrics \code{\link{accu}} in the plot.
Default: \code{show.accu = TRUE}.}

\item{w.acc}{Weigthing parameter \code{w} used to compute
weighted accuracy \code{w.acc} in \code{\link{comp_accu}}.
Default: \code{w.acc = .50}.}

\item{title.lbl}{Text label for current plot title.
Default: \code{title.lbl = txt$scen.lbl}.}

\item{col.pal}{Color palette.
Default: \code{col.pal = pal} (see \code{\link{pal}} and \code{\link{init_pal}}).}
}
\description{
\code{plot_bar} draws bar charts that
represent the proportions of frequencies in the current
population \code{\link{popu}} as relatives sizes of
rectangular areas.
}
\details{
If a sufficient and valid set of 3 essential probabilities
(\code{\link{prev}}, and
\code{\link{sens}} or its complement \code{\link{mirt}}, and
\code{\link{spec}} or its complement \code{\link{fart}})
is provided, new frequency information \code{\link{freq}}
and a new population table \code{\link{popu}}
are computed from scratch.  Otherwise, the existing
population \code{\link{popu}} is shown.

Rectangles corresponding to the areas of the mosaic plot
can be visualized byopting for vertical rectangles (by selecting
the option \code{box = "vr"}) in \code{\link{plot_tree}}
and \code{\link{plot_fnet}}.
}
\examples{

# Basics:
plot_bar()  # => plot with default options
plot_bar(prev = .33, sens = .75, spec = .66, title.lbl = "Test 1")
plot_bar(N = 1000, prev = .33, sens = .75, spec = .66, title.lbl = "Test 2", show.freq = TRUE)

# Note:
plot_bar(N = 3, prev = .33, sens = .75, spec = .66,
         title.lbl = "Test rounding effects",
         show.freq = TRUE)  # => Rounding of freq, but not of proportions shown in plot.


}
\seealso{
\code{\link{comp_popu}} computes the current population;
\code{\link{popu}} contains the current population;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{freq}} contains current frequency information;
\code{\link{num}} for basic numeric parameters;
\code{\link{txt}} for current text settings;
\code{\link{pal}} for current color settings

Other visualization functions: \code{\link{plot.riskyr}},
  \code{\link{plot_curve}}, \code{\link{plot_fnet}},
  \code{\link{plot_icons}}, \code{\link{plot_mosaic_2}},
  \code{\link{plot_mosaic}}, \code{\link{plot_plane}},
  \code{\link{plot_tree}}
}
