% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crisk.R
\name{plot_crisk}
\alias{plot_crisk}
\title{Plot a cumulative risk curve.}
\usage{
plot_crisk(
  x,
  y,
  x_from = NA,
  x_to = NA,
  fit_curve = FALSE,
  show_pas = TRUE,
  show_rem = TRUE,
  show_aux = TRUE,
  show_num = TRUE,
  show_inc = FALSE,
  show_grid = FALSE,
  col_pal = pal_crisk,
  arr_c = -3,
  title_lbl = "Cumulative risk",
  x_lbl = "Age (in years)",
  y_lbl = "Population risk",
  y2_lbl = "",
  cex_lbl = 0.95,
  mar_notes = TRUE,
  ...
)
}
\arguments{
\item{x}{Values on an x-dimension on which risk is expressed
(required, as a vector).}

\item{y}{Values of cumulative risks on an y-dimension
(required, as a vector),
as monotonically increasing percentage values
(0 <= y <= 100).
Pairs of \code{x} and \code{y} are assumed to
correspond to each other.}

\item{x_from}{Start value of risk increment.}

\item{x_to}{End value of risk increment.}

\item{fit_curve}{Boolean: Fit a curve to \code{x}-\code{y}-data?
Default: \code{fit_curve = FALSE}.}

\item{show_pas}{Boolean: Show past/passed risk?
Default: \code{show_pas = TRUE}.}

\item{show_rem}{Boolean: Show remaining risk?
Default: \code{show_rem = TRUE}.}

\item{show_aux}{Boolean: Show auxiliary elements
(i.e., lines, points, and text labels)?
Default: \code{show_aux = TRUE}.}

\item{show_num}{Boolean: Show numeric values,
provided that \code{show_aux = TRUE}.
Default: \code{show_num = TRUE}.}

\item{show_inc}{Boolean: Show risk increments?
Default: \code{show_inc = FALSE}.}

\item{show_grid}{Boolean: Show grid lines?
Default: \code{show_grid = FALSE}.}

\item{col_pal}{Color palette (as a named vector).
Default: \code{col_pal = pal_crisk}.}

\item{arr_c}{Arrow code for symbols at ends of probability links
(as a numeric value \code{-3 <= arr_c <= +6}),
with the following options:
  \itemize{
  \item \code{-1} to \code{-3}: points at one/other/both end/s;
  \item \code{0}: no symbols;
  \item \code{+1} to \code{+3}: V-arrow at one/other/both end/s;
  \item \code{+4} to \code{+6}: T-arrow at one/other/both end/s.
}
Default: \code{arr_c = -3} (points at both ends).}

\item{title_lbl}{Text label for current plot title.
Default: \code{title_lbl = "Cumulative risk"}.}

\item{x_lbl}{Text label of x-axis (at bottom).
Default: \code{x_lbl = "Age (in years)"}.}

\item{y_lbl}{Text label of y-axis (on left).
Default: \code{y_lbl = "Population risk"}.}

\item{y2_lbl}{Text label of 2nd y-axis (on right).
Default: \code{y2_lbl = ""} (formerly "Remaining risk").}

\item{cex_lbl}{Scaling factor for text labels.
Default: \code{cex_lbl = .95}.}

\item{mar_notes}{Boolean option for showing margin notes.
Default: \code{mar_notes = TRUE}.}

\item{...}{Other (graphical) parameters.}
}
\value{
Nothing (NULL).
}
\description{
\code{plot_crisk} creates visualizations of
cumulative risks.
}
\details{
Assumes data inputs \code{x} and \code{y},
which correspond to each other so that
\code{y} is a (monotonically increasing)
probability density function (over risk increments
expressed as a function of \code{x}).
}
\examples{
# Data:
x <- seq(0, 100, by = 10)
y <- c(0, 0, 0, 10, 24, 50, 72, 80, 83, 85, 85)

plot_crisk(x, y)
plot_crisk(x, y, x_from = 40)
plot_crisk(x, y, x_from = 40, x_to = 60)
plot_crisk(x, y, fit_curve = FALSE, title = "My title", mar_notes = TRUE)
plot_crisk(x, y, x_from = 40, x_to = 60)  # provided points
plot_crisk(x, y, x_from = 46, x_to = 76)  # predicted points

# Small x- and y-values and linear increases:
plot_crisk(x = 1:10, y = seq( 1, 20, by = 2), x_from = 4,   x_to = 8, show_inc = TRUE)
plot_crisk(x = 2:10, y = seq(12, 28, by = 2), x_from = 4.5, x_to = 8.5, show_inc = TRUE)

# Versions:
plot_crisk(x, y, x_from = 42, x_to = 62, show_rem = FALSE, show_aux = FALSE)
plot_crisk(x, y, x_from = 42, x_to = 62, show_pas = FALSE, show_aux = FALSE)
plot_crisk(x, y, x_from = 42, x_to = 62, show_aux = FALSE)
plot_crisk(x, y, x_from = 42, x_to = 62, show_num = FALSE)
plot_crisk(x, y, x_from = 42, x_to = 62, show_pas = FALSE, show_rem = FALSE)

# Note: Showing all may overplot/overwhelm:
plot_crisk(x, y, x_from = 48, x_to = 68, show_inc = TRUE, show_grid = TRUE)

}
\seealso{
\code{\link{pal_crisk}} corresponding color palette.

Other visualization functions: 
\code{\link{plot.riskyr}()},
\code{\link{plot_area}()},
\code{\link{plot_bar}()},
\code{\link{plot_curve}()},
\code{\link{plot_fnet}()},
\code{\link{plot_icons}()},
\code{\link{plot_mosaic}()},
\code{\link{plot_plane}()},
\code{\link{plot_prism}()},
\code{\link{plot_tab}()},
\code{\link{plot_tree}()}
}
\concept{visualization functions}
