% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_risk.R
\name{plot_cum_bar}
\alias{plot_cum_bar}
\title{Plot cumulative risk dynamics (as bar chart)}
\usage{
plot_cum_bar(
  r = 0.5,
  t = 1,
  N = 100,
  horizontal = TRUE,
  sort = FALSE,
  N_max = 100,
  bar_width = 0.5,
  show_trans = TRUE,
  show_ev = TRUE,
  show_n = FALSE
)
}
\arguments{
\item{r}{risk (probability of occurrence per time period)}

\item{t}{time periods/rounds.}

\item{N}{population size.
Default: \code{N = 100} expresses risks as percentages,
\code{N = 1} as probabilities, else frequencies.}

\item{horizontal}{logical: Draw horizontal vs. vertical bars?}

\item{sort}{logical: Sort outputs by number of event occurrences?
Default: \code{sort = FALSE}.}

\item{N_max}{maximum N value plotted (for zooming in for small \code{r} values).
Default value should be set to population size \code{N}.}

\item{bar_width}{width of (horizontal/vertical) bar per time period.
Default: \code{bar_width = .50}.}

\item{show_trans}{logical: Show transition polygons (between bars)?
Default: \code{show_trans = TRUE}.}

\item{show_ev}{logical: Show number of risky event occurrence (as bar label)?
Default: \code{show_ev = TRUE}.}

\item{show_n}{logical: Show population frequency of risky event occurrences (as bar label)?
Default: \code{show_n = FALSE}.}
}
\value{
data of p-values, named by number of event occurrences
(invisibly, as list of named vectors, for each time period t).
}
\description{
plot_cum_bar plots the results of cumulative risk dynamics
as a bar chart (with percentages of risk event counts
for each period t on a horizontal bar).
}
