% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_risk.R
\name{plot_cum_bar}
\alias{plot_cum_bar}
\title{Plot cumulative risk dynamics}
\usage{
plot_cum_bar(r = 1/2, t = 1, N = 100, sort = FALSE)
}
\arguments{
\item{N}{population size.}

\item{data}{The data to plot (as list of named probabilities).}
}
\description{
plot_cum_bar plots the results of cumulative risk dynamics
as a barchart (with percentages of risk event counts
for each period t on a horizontal bar).
}
