% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_risk.R
\name{plot_cum_bar}
\alias{plot_cum_bar}
\title{Plot cumulative risk dynamics}
\usage{
plot_cum_bar(r = 1/2, t = 1, N = 100, sort = FALSE, x_max = 100)
}
\arguments{
\item{r}{risk (per time period)}

\item{t}{number of time periods}

\item{N}{population size.
Default: \code{N = 100} expresses risks as percentages,
\code{N = 1} as probabilities, else frequencies.}

\item{sort}{logical: sort outputs (by number of event occurrences)?}

\item{x_max}{maximum x-value (for zooming).
Default should equal population size \code{N}.}
}
\description{
plot_cum_bar plots the results of cumulative risk dynamics
as a barchart (with percentages of risk event counts
for each period t on a horizontal bar).
}
