% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_risk.R
\name{plot_cum_bar}
\alias{plot_cum_bar}
\title{Plot cumulative risk dynamics}
\usage{
plot_cum_bar(
  r = 1/2,
  t = 1,
  N = 100,
  horizontal = TRUE,
  sort = FALSE,
  N_max = 100
)
}
\arguments{
\item{r}{risk (probability of occurrence per time period)}

\item{t}{number of time periods}

\item{N}{population size.
Default: \code{N = 100} expresses risks as percentages,
\code{N = 1} as probabilities, else frequencies.}

\item{horizontal}{logical: Draw horizontal vs. vertical bars?}

\item{sort}{logical: sort outputs (by number of event occurrences)?}

\item{N_max}{maximum N value (for zooming in for small \code{r} values).
Default value should be set to population size \code{N}.}
}
\description{
plot_cum_bar plots the results of cumulative risk dynamics
as a barchart (with percentages of risk event counts
for each period t on a horizontal bar).
}
