% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_popu.R
\docType{data}
\name{popu}
\alias{popu}
\title{A population table based on current frequencies.}
\format{An object of class \code{data.frame} with 12 rows and 3 columns.}
\usage{
popu
}
\value{
A data frame \code{popu}
containing \code{\link{N}} rows (individual cases)
and 3 columns (\code{"Truth", "Decision", "SDT"})
encoded as ordered factors
(with 2, 2, and 4 levels, respectively).
}
\description{
\code{popu} is an R data frame that is computed
by \code{\link{comp_popu}} from the current
frequency information (contained in \code{\link{freq}}).
Each individual is represented as a row;
columns represent the individual's
condition (\code{TRUE} or \code{FALSE}),
a corresponding decision
(also encoded as \code{TRUE} = positive or \code{FALSE} = negative),
and its classification (in SDT terms) as either
true positive (an individual hit),
false negative (an individual miss),
false positive (an individual false alarm), or
true negative (an individual correct rejection).
}
\details{
\code{comp_popu} uses the current text information
contained in \code{\link{txt}} to define
the labels of conditions, decisions, and
SDT classifications.

A visualization of the current population
\code{popu} is provided by \code{\link{plot_iconarray}}.
}
\examples{
popu <- comp_popu()  # => initializes popu with current values of freq and txt
dim(popu)            # => N x 3
head(popu)           # => shows head of data frame

}
\seealso{
The corresponding generating function \code{\link{comp_popu}};
\code{\link{num}} for basic numeric parameters;
\code{\link{freq}} for current frequency information;
\code{\link{txt}} for current text settings.
}
\keyword{datasets}
