% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prob.R
\docType{data}
\name{PPV}
\alias{PPV}
\alias{precision}
\title{The positive predictive value of a decision process or diagnostic procedure.}
\format{An object of class \code{logical} of length 1.}
\usage{
PPV
}
\description{
\code{PPV} defines some decision's positive predictive value (PPV):
The conditional probability of the condition being \code{TRUE}
provided that the decision is positive.
}
\details{
Understanding or obtaining the positive predictive value \code{PPV}:

\itemize{

\item Definition:
  \code{PPV} is the conditional probability
  for the condition being \code{TRUE}
  given a positive decision:

\code{PPV = p(condition = TRUE | decision = positive)}

or the probability of a positive decision being correct.

\item Alternative names:
  \code{precision}

\item Relationships:

a. \code{PPV} is the complement of the
  false discovery or false detection rate \code{\link{FDR}}:

\code{PPV = 1 - FDR}

b. \code{PPV} is the opposite conditional probability
  -- but not the complement --
  of the sensitivity \code{\link{sens}}:

\code{sens = p(decision = positive | condition = TRUE)}

\item In natural frequencies,
  \code{PPV} is the ratio of individuals for which
  \code{decision = positive} and \code{condition = TRUE}
  divided by the number of all individuals for which
  \code{decision = positive}:

\code{PPV = n(decision = positive & condition = TRUE) / n(decision = positive)}

\code{PPV = n.hi / (n.hi + n.fa) }

\item \code{PPV} is a feature of a decision process
  or diagnostic procedure and
  -- similar to the sensitivity \code{\link{sens}} --
  a measure of correct decisions (positive decisions
  that are actually TRUE).

However, due to being a conditional probability,
  the value of \code{PPV} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
PPV <- .55     # => sets a positive predictive value of 55\%
PPV <- 55/100  # => (condition = TRUE) for 55 people out of 100 people for which (decision = positive)
is_prob(PPV)   # => TRUE (as PPV is a probability)

}
\references{
Consult \href{https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values}{Wikipedia} for additional information.
}
\seealso{
\code{\link{num}} contains basic numeric variables;
\code{\link{comp_PPV}} and \code{\link{comp_PPV_freq}} compute PPVs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{is_prob}} verifies probability inputs;
\code{\link{comp_prob}} computes derived probabilities;
\code{\link{comp_freq}} computes natural frequencies from probabilities

Other functions computing probabilities: \code{\link{FDR}},
  \code{\link{FOR}}, \code{\link{NPV}},
  \code{\link{comp_complement}}, \code{\link{comp_fart}},
  \code{\link{comp_prob}}, \code{\link{comp_spec}}
}
\keyword{datasets}
