% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskyr_class.R
\name{print.summary.riskyr}
\alias{print.summary.riskyr}
\title{Printing summarized risk information.}
\format{Printed output of a "summary.riskyr" object.}
\usage{
\method{print}{summary.riskyr}(x = NULL, ...)
}
\arguments{
\item{x}{An object of class "summary.riskyr", usually a result of a call to \code{summary.riskyr}.}

\item{...}{Additional parameters (to be passed to generic print function).}
}
\description{
\code{print.summary.riskyr} provides a \code{print} method for objects of class "summary.riskyr".
}
\examples{
summary(scenarios$n4)

}
\concept{print functions}
