% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_num_prob.R
\docType{data}
\name{prob}
\alias{prob}
\title{List current probability information.}
\format{An object of class \code{list} of length 8.}
\usage{
prob
}
\description{
\code{prob} is a list of named numeric variables
containing 4 basic (1 non-conditional and 3 conditional) probabilities
and 4 derived (conditional) probabilities:
}
\details{
\enumerate{

\item the condition's prevalence value \code{\link{prev}}
 (i.e., the probability of condition being \code{TRUE}).

\item the decision's sensitivity value \code{\link{sens}}
 (i.e., the conditional probability of a positive decision
 provided that the condition is \code{TRUE}).

\item the decision's specificity value \code{\link{spec}}
 (i.e., the conditional probability
 of a negative decision provided that the condition is \code{FALSE}).

\item the decision's false alarm rate \code{\link{fart}}
 (i.e., the conditional probability
 of a positive decision provided that the condition is \code{FALSE}).

\item the decision's positive predictive value \code{\link{PPV}}
(i.e., the conditional probability of the condition being \code{TRUE}
provided that the decision is positive)

\item the decision's negative predictive value \code{\link{NPV}}
(i.e., the conditional probability of the condition being \code{FALSE}
provided that the decision is negative)

\item the decision's false discovery or false detection rate \code{\link{FDR}}
(i.e., the conditional probability of the condition being \code{FALSE}
provided that the decision is positive)

\item the decision's false omission rate \code{\link{FOR}}
(i.e., the conditional probability of the condition being \code{TRUE}
provided that the decision is negative)

}

These probabilities are computed from basic probabilities
(contained in \code{\link{num}}) and computed by using
\code{\link{comp_prob}}.

The list \code{prob} is the probability counterpart
to the list containing frequency information \code{\link{freq}}.

Note that inputs of extreme probabilities (of 0 or 1)
may yield unexpected values (e.g., an \code{\link{NPV}}
value of NaN when \code{\link{is_perfect}}
evaluates to \code{TRUE}).
}
\examples{
prob <- comp_prob()  # => initialize prob to default parameters
prob                 # => show current values
length(prob)         # => 8

}
\seealso{
\code{\link{comp_freq}} computes current frequency information;
\code{\link{num}} contains basic numeric variables;
\code{\link{init_num}} initializes basic numeric variables;
\code{\link{prob}} contains current probability information

Other lists containing scenario settings: \code{\link{freq}},
  \code{\link{num}}
}
\keyword{datasets}
