% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskyr_class.R
\name{riskyr}
\alias{riskyr}
\title{Create riskyr scenarios.}
\format{An object of class "riskyr"
with 21 entries of textual and numeric information describing a risk-related scenario.}
\usage{
riskyr(scen_lbl = "", scen_lng = txt$scen_lng,
  scen_txt = txt$scen_txt, popu_lbl = txt$popu_lbl,
  cond_lbl = txt$cond_lbl, cond.true_lbl = txt$cond.true_lbl,
  cond.false_lbl = txt$cond.false_lbl, dec_lbl = txt$dec_lbl,
  dec.pos_lbl = txt$dec.pos_lbl, dec.neg_lbl = txt$dec.neg_lbl,
  sdt_lbl = txt$sdt_lbl, hi_lbl = txt$hi_lbl, mi_lbl = txt$mi_lbl,
  fa_lbl = txt$fa_lbl, cr_lbl = txt$cr_lbl, prev = NA, sens = NA,
  spec = NA, fart = NA, N = NA, hi = NA, mi = NA, fa = NA,
  cr = NA, scen_src = txt$scen_src, scen_apa = txt$scen_apa)
}
\arguments{
\item{scen_lbl}{The current scenario title (sometimes in Title Caps).}

\item{scen_lng}{Language of the current scenario (as character code).
Options: \code{"en"} for English, \code{"de"} for  German.}

\item{scen_txt}{A longer text description of the current scenario
(which may extend over several lines).}

\item{popu_lbl}{A brief description of the current target population \code{\link{popu}} or sample.}

\item{cond_lbl}{A name for the \emph{condition} or feature (e.g., some disease) currently considered.}

\item{cond.true_lbl}{A label for the \emph{presence} of the current condition
or \code{\link{cond.true}} cases (the condition's true state of TRUE).}

\item{cond.false_lbl}{A label for the \emph{absence} of the current condition
or \code{\link{cond.false}} cases (the condition's true state of FALSE).}

\item{dec_lbl}{A name for the \emph{decision} or judgment (e.g., some diagnostic test) currently made.}

\item{dec.pos_lbl}{A label for \emph{positive} decisions
or \code{\link{dec.pos}} cases (e.g., predicting the presence of the condition).}

\item{dec.neg_lbl}{A label for \emph{negative} decisions
or \code{\link{dec.neg}} cases (e.g., predicting the absence of the condition).}

\item{sdt_lbl}{A name for the combination of \emph{condition} and \emph{decision} currently made.}

\item{hi_lbl}{A label for \emph{hits} or \emph{true positives} \code{\link{hi}}
(i.e., correct decisions of the presence of the condition, when the condition is actually present).}

\item{mi_lbl}{A label for \emph{misses} or \emph{false negatives} \code{\link{mi}}
(i.e., incorrect decisions of the absence of the condition when the condition is actually present).}

\item{fa_lbl}{A label for \emph{false alarms} or \emph{false positives} \code{\link{fa}}
(i.e., incorrect decisions of the presence of the condition when the condition is actually absent).}

\item{cr_lbl}{A label for \emph{correct rejections} or \emph{true negatives} \code{\link{cr}}
(i.e., a correct decision of the absence of the condition, when the condition is actually absent).

Numeric elements:}

\item{prev}{The condition's prevalence \code{\link{prev}}
(i.e., the probability of condition being \code{TRUE}).}

\item{sens}{The decision's sensitivity \code{\link{sens}}
(i.e., the conditional probability of a positive decision
provided that the condition is \code{TRUE}).
\code{sens} is optional when its complement \code{mirt} is provided.}

\item{spec}{The decision's specificity value \code{\link{spec}}
(i.e., the conditional probability
of a negative decision provided that the condition is \code{FALSE}).
\code{spec} is optional when its complement \code{fart} is provided.}

\item{fart}{The decision's false alarm rate \code{\link{fart}}
(i.e., the conditional probability
of a positive decision provided that the condition is \code{FALSE}).
\code{fart} is optional when its complement \code{spec} is provided.}

\item{N}{The number of individuals in the scenario's population.
A suitable value of \code{\link{N}} is computed, if not provided.}

\item{hi}{The number of hits \code{\link{hi}} (or true positives).}

\item{mi}{The number of misses \code{\link{mi}} (or false negatives).}

\item{fa}{The number of false alarms \code{\link{fa}} (or false positives).}

\item{cr}{The number of correct rejections \code{\link{cr}} (or true negatives).

Source information:}

\item{scen_src}{Source information for the current scenario.}

\item{scen_apa}{Source information for the current scenario
in the style of the American Psychological Association (APA style).}
}
\value{
An object of class "riskyr" describing a risk-related scenario.

Text elements (all elements of \code{\link{txt}}:
}
\description{
The instantiation function \code{riskyr} is used to create
scenarios of class "riskyr",
which can then be visualized by the \code{plot} method \code{\link{plot.riskyr}}
and summarized by the \code{summary} method \code{\link{summary.riskyr}}.
}
\details{
Beyond basic scenario information
only the population size \code{\link{N}} and the essential probabilities
\code{\link{prev}}, \code{\link{sens}}, \code{\link{spec}}, and \code{\link{fart}}
are used and returned.
}
\examples{
# Defining a scenario:
scen.reoffend <- riskyr(scen_lbl = "Identify reoffenders",
                        cond_lbl = "being a reoffender",
                        popu_lbl = "Prisoners",
                        cond.true_lbl = "has reoffended",
                        cond.false_lbl = "has not reoffended",
                        dec_lbl = "test result",
                        dec.pos_lbl = "will reoffend",
                        dec.neg_lbl = "will not reoffend",
                        sdt_lbl = "combination",
                        hi_lbl = "reoffender found", mi_lbl = "reoffender missed",
                        fa_lbl = "false accusation", cr_lbl = "correct release",
                        prev = .45,  # prevalence of being a reoffender.
                        sens = .98,
                        spec = .46, fart = NA,  # (provide 1 of 2)
                        N = 753,
                        scen_src = "Example scenario")

# Using a scenario:
summary(scen.reoffend)
plot(scen.reoffend)

# 2 ways of defining the same scenario:
s1 <- riskyr(prev = .5, sens = .5, spec = .5, N = 100)  # s1: define by 3 prob & N
s2 <- riskyr(hi = 25, mi = 25, fa = 25, cr = 25)        # s2: same scenario by 4 freq
all.equal(s1, s2)  # should be TRUE

# Ways to work:
riskyr(prev = .5, sens = .5, spec = .5, hi = 25, mi = 25, fa = 25, cr = 25)  # works (consistent)
riskyr(prev = .5, sens = .5, spec = .5, hi = 25, mi = 25, fa = 25)           # works (ignores freq)

## Watch out for:
# riskyr(hi = 25, mi = 25, fa = 25, cr = 25, N = 101)  # warns, uses actual sum of freq
# riskyr(prev = .4, sens = .5, spec = .5, hi = 25, mi = 25, fa = 25, cr = 25)  # warns, uses freq

}
\seealso{
Other riskyr scenario functions: \code{\link{plot.riskyr}},
  \code{\link{read_popu}}, \code{\link{summary.riskyr}}
}
\concept{riskyr scenario functions}
