% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prob.R
\docType{data}
\name{spec}
\alias{spec}
\alias{TNR}
\title{The specificity of a decision process or diagnostic procedure.}
\format{An object of class \code{numeric} of length 1.}
\usage{
spec
}
\description{
\code{spec} defines a decision's specificity value (or correct rejection rate):
The conditional probability of the decision being negative
if the condition is FALSE.
}
\details{
Understanding or obtaining the specificity value \code{spec}:

\itemize{

\item Definition:
  \code{spec} is the conditional probability
  for a (correct) negative decision given that
  the condition is \code{FALSE}:

\code{spec = p(decision = negative | condition = FALSE)}

or the probability of correctly detecting false cases
  (\code{condition = FALSE}).

\item Alternative names:
  true negative rate (\code{TNR}), correct rejection rate

\item Relationships:

a. \code{spec} is the complement of the
  false alarm rate \code{\link{fart}}:

\code{spec = 1 - fart}

b. \code{spec} is the opposite conditional probability
  -- but not the complement --
  of the negative predictive value \code{\link{NPV}}:

\code{NPV = p(condition = FALSE | decision = negative)}

\item In natural frequencies,
  \code{spec} is the ratio of individuals for which
  \code{decision = negative} divided by the number of
  individuals for which \code{condition = FALSE}:

\code{spec = n(decision = negative) / n(condition = FALSE)}

\item \code{spec} is a feature of a decision process
  or diagnostic procedure and a measure of
  correct decisions (true negatives).

However, due to being a conditional probability,
  the value of \code{spec} also depends on the
  condition's prevalence value \code{\link{prev}}.

}
}
\examples{
spec <- .75     # => sets a specificity value of 75\%
spec <- 75/100  # => (decision = negative) for 75 people out of 100 people for which (condition = FALSE)
is_prob(spec)   # => TRUE (as spec is a probability)
}
\references{
Consult \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{Wikipedia} for additional information.
}
\seealso{
\code{\link{comp_spec}} computes \code{spec} as the complement of \code{\link{fart}};
\code{\link{prob}} contains current probability information;
\code{\link{comp_prob}} computes current probability information;
\code{\link{num}} contains basic numeric parameters;
\code{\link{init_num}} initializes basic numeric parameters;
\code{\link{comp_freq}} computes current frequency information;
\code{\link{is_prob}} verifies probability inputs.

Other basic probabilities: \code{\link{fart}},
  \code{\link{mirt}}, \code{\link{prev}},
  \code{\link{sens}}
}
\keyword{datasets}
