% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_txt.R
\docType{data}
\name{txt_TF}
\alias{txt_TF}
\title{Alternative text labels (TP, FN, FP, TN).}
\format{An object of class \code{list} of length 21.}
\usage{
txt_TF
}
\description{
\code{txt_TF} is initialized to alternative text labels
to define a frequency naming scheme in which
(hi, mi, fa, cr) are called (TP, FN, FP, TN).
}
\details{
See \code{\link{txt}} for details and default text information.

Assign \code{txt <- txt_TF} to use as default text labels.
}
\examples{
txt_TF        # shows text labels of txt_TF
txt_TF["hi"]  # shows the current label for hits ("TP")
txt_TF["hi"] <- "hit" # defines a new label for hits (true positives, TP)

}
\seealso{
\code{\link{txt}} contains current text information;
\code{\link{init_txt}} initializes text information;
\code{\link{pal}} contains current color information;
\code{\link{init_pal}} initializes color information.

Other lists containing current scenario information: \code{\link{accu}},
  \code{\link{freq}}, \code{\link{num}},
  \code{\link{pal_bwp}}, \code{\link{pal_bw}},
  \code{\link{pal_kn}}, \code{\link{pal_mbw}},
  \code{\link{pal_mod}}, \code{\link{pal_org}},
  \code{\link{pal_rgb}}, \code{\link{pal_vir}},
  \code{\link{pal}}, \code{\link{prob}},
  \code{\link{txt_org}}, \code{\link{txt}}
}
\concept{lists containing current scenario information}
\keyword{datasets}
